### 2019-02-19 #1267 @whedon
[PRE REVIEW]: NiBetaSeries: task related correlations in fMRI
**Submitting author:** @jdkent (<a href="http://orcid.org/0000-0002-4892-2659">James Kent</a>)
**Repository:** <a href="https://github.com/HBClab/NiBetaSeries" target ="_blank">https://github.com/HBClab/NiBetaSeries</a>
**Version:** v0.2.3
**Reviewers:** snastase
