### 2019-02-27 #1267 @jdkent
Thanks @arokem!
Here are a few (in no particular order):
mnarayan
snastase
effigies
