### 2019-02-28 #1267 @arokem
Hey @snastase : would you be willing to review this paper for JOSS?
