### 2019-03-04 #1295 @whedon
[REVIEW]: NiBetaSeries: task related correlations in fMRI
**Submitting author:** @jdkent (<a href="http://orcid.org/0000-0002-4892-2659">James Kent</a>)
**Repository:** <a href="https://github.com/HBClab/NiBetaSeries" target ="_blank">https://github.com/HBClab/NiBetaSeries</a>
**Version:** v0.3.2
