### 2019-04-12 HBClab/NiBetaSeries#127 @jdkent
make docker/singularity container
[original comment](https://github.com/openjournals/joss-reviews/issues/1295#issuecomment-477289593)
> Would it be worth making a Singularity image for this? For example, I'm almost exclusively running fMRIPRep and MRIQC apps on a server via Singularity because I don't have installation privileges. I suppose the alternative is indicating to users to pip install nibetaseries in a conda environment or something along these lines.
