### 2019-04-12 HBClab/NiBetaSeries#128 @jdkent
make it clear only python3 is usable
[original comment](https://github.com/openjournals/joss-reviews/issues/1295#issuecomment-477289593)

> Speaking of installation, if you're degenerate like me and still have a Python 2.7 installation on your machine, pip install nibetaseries will try to install this in 2.7 and wreck. Something like pip3 install nibetaseries or python3 -m pip install nibetaseries works more reliably. I would slightly expand the installation documentation page, specifying Python version, etc.
