### 2019-04-12 HBClab/NiBetaSeries#129 @jdkent
add binder link
[original comment](https://github.com/openjournals/joss-reviews/issues/1295#issuecomment-477289593)

> It might be worth pointing users to Binder (https://mybinder.org) for running the tutorial Jupyter Notebook interactively.
