### 2019-04-14 HBClab/NiBetaSeries#127 @PeerHerholz
If okay for you @jdkent and everybody else, I would tackle this, generating a `bash` script that utilizes [neurodocker](https://github.com/kaczmarj/neurodocker) to generate a Docker- and Singularityfile. I would open a respective branch on my fork and push to master, including a `README.md` that has a small stepwise tutorial on how to recreate the images. Regarding `hosting` on `docker hub` I would suggest that you @jdkent, build the images locally and then push it to your `docker hub`, eh? Once that is done, we can easily do the binder related fun addressed in #129.
