### 2019-04-15 HBClab/NiBetaSeries#127 @jdkent
Thanks @PeerHerholz! That sounds like a reasonable plan of action. For now, all dependencies should be installable via pip, but this may change in the future.

One stretch goal would be to also automate the building of the container (probably through `circleci`) in an analogous way that [fmriprep](https://github.com/poldracklab/fmriprep/blob/d29902e2a1e03840dc0c0e55e9dedd0d2dc30fca/.circleci/config.yml#L4-L76) builds containers.
