### 2019-04-15 b73821f
initial draft of docker / singularity image
- this PR includes the initial draft of a NiBetaSeries docker and/or singularity image
- it includes both, a Docker- and Singularityfile (`Dockerfile`, `Singularity`), as well as a bash script (`generate_nibetaseries_images.sh`) that recreates both images and a small `README.md`
- an initial test of the Docker image performance was done by running the [plot_run_nibetaseries example](https://github.com/HBClab/NiBetaSeries/blob/master/examples/plot_run_nibetaseries.py) within the Docker image, the respective gist can be found [here](https://gist.github
[…]
