+++ docs/auto_examples/sg_execution_times.rst
@@ -1,0 +1,10 @@
+
+:orphan:
+
+.. _sphx_glr_auto_examples_sg_execution_times:
+
+Computation times
+=================
+**02:33.297** total execution time for **auto_examples** files:
+
+- **02:33.297**: :ref:`sphx_glr_auto_examples_plot_run_nibetaseries.py` (``plot_run_nibetaseries.py``)
