+++ docs/auto_examples/plot_run_nibetaseries.rst
@@ -1,0 +1,1430 @@
+.. note::
+    :class: sphx-glr-download-link-note
+
+    Click :ref:`here <sphx_glr_download_auto_examples_plot_run_nibetaseries.py>` to download the full example code
+.. rst-class:: sphx-glr-example-title
+
+.. _sphx_glr_auto_examples_plot_run_nibetaseries.py:
+
+
+Running NiBetaSeries using ds000164 (Stroop Task)
+===============================================================
+
+This example runs through a basic call of NiBetaSeries using
+the commandline entry point ``nibs``.
+While this example is using python, typically ``nibs`` will be
+called directly on the commandline.
+
+Import all the necessary packages
+=================================
+
+
+.. code-block:: default
+
+
+    import tempfile  # make a temporary directory for files
+    import os  # interact with the filesystem
+    import urllib.request  # grad data from internet
+    import tarfile  # extract files from tar
+    from subprocess import Popen, PIPE, STDOUT  # enable calling commandline
+
+    import matplotlib.pyplot as plt  # manipulate figures
+    import seaborn as sns  # display results
+    import pandas as pd   # manipulate tabular data
+
+
+
+
+
+
+
+Download relevant data from ds000164 (and Atlas Files)
+======================================================
+The subject data came from `openneuro <https://openneuro.org/datasets/ds000164/versions/00001/>`_.
+The atlas data came from a `recently published parcellation <https://www.ncbi.nlm.nih.gov/pubmed/28981612>`_
+in a publically accessible github repository.
+
+
+.. code-block:: default
+
+
+    # atlas github repo for reference:
+    """https://github.com/ThomasYeoLab/CBIG/raw/master/stable_projects/\
+    brain_parcellation/Schaefer2018_LocalGlobal/Parcellations/MNI/"""
+    data_dir = tempfile.mkdtemp()
+    print('Our working directory: {}'.format(data_dir))
+
+    # download the tar data
+    url = "https://www.dropbox.com/s/qoqbiya1ou7vi78/ds000164-test_v1.tar.gz?dl=1"
+    tar_file = os.path.join(data_dir, "ds000164.tar.gz")
+    u = urllib.request.urlopen(url)
+    data = u.read()
+    u.close()
+
+    # write tar data to file
+    with open(tar_file, "wb") as f:
+        f.write(data)
+
+    # extract the data
+    tar = tarfile.open(tar_file, mode='r|gz')
+    tar.extractall(path=data_dir)
+
+    os.remove(tar_file)
+
+
+
+
+
+.. rst-class:: sphx-glr-script-out
+
+ Out:
+
+ .. code-block:: none
+
+    Our working directory: /var/folders/7w/skj0rj1d4hgdhxl6813xkc6h0000gn/T/tmppvvwl83a
+
+
+Display the minimal dataset necessary to run nibs
+=================================================
+
+
+.. code-block:: default
+
+
+
+    # https://stackoverflow.com/questions/9727673/list-directory-tree-structure-in-python
+    def list_files(startpath):
+        for root, dirs, files in os.walk(startpath):
+            level = root.replace(startpath, '').count(os.sep)
+            indent = ' ' * 4 * (level)
+            print('{}{}/'.format(indent, os.path.basename(root)))
+            subindent = ' ' * 4 * (level + 1)
+            for f in files:
[…]
