+++ docs/auto_examples/index.rst
@@ -1,0 +1,60 @@
+:orphan:
+
+
+
+.. _sphx_glr_auto_examples:
+
+How to run NiBetaSeries
+=======================
+
+Below is an example run through of NiBetaSeries
+
+
+
+.. raw:: html
+
+    <div class="sphx-glr-thumbcontainer" tooltip="This example runs through a basic call of NiBetaSeries using the commandline entry point ``nibs...">
+
+.. only:: html
+
+    .. figure:: /auto_examples/images/thumb/sphx_glr_plot_run_nibetaseries_thumb.png
+
+        :ref:`sphx_glr_auto_examples_plot_run_nibetaseries.py`
+
+.. raw:: html
+
+    </div>
+
+
+.. toctree::
+   :hidden:
+
+   /auto_examples/plot_run_nibetaseries
+.. raw:: html
+
+    <div style='clear:both'></div>
+
+
+
+.. only :: html
+
+ .. container:: sphx-glr-footer
+    :class: sphx-glr-footer-gallery
+
+
+  .. container:: sphx-glr-download
+
+    :download:`Download all examples in Python source code: auto_examples_python.zip <//Users/peerherholz/google_drive/GitHub/NiBetaSeries/docs/auto_examples/auto_examples_python.zip>`
+
+
+
+  .. container:: sphx-glr-download
+
+    :download:`Download all examples in Jupyter notebooks: auto_examples_jupyter.zip <//Users/peerherholz/google_drive/GitHub/NiBetaSeries/docs/auto_examples/auto_examples_jupyter.zip>`
+
+
+.. only:: html
+
+ .. rst-class:: sphx-glr-signature
+
+    `Gallery generated by Sphinx-Gallery <https://sphinx-gallery.readthedocs.io>`_
