### 2019-04-16 HBClab/NiBetaSeries#127 @PeerHerholz
The initial draft of the images is proposed in #140.
