### 2019-04-17 0525f66
revert commit 4e1fa4a
- `auto_examples` and `nibetaseries_work` were committed in the last version of the PR, but shouldn't
- hence, the respective changes were reverted
