+++ docs/auto_examples/plot_run_nibetaseries.py
@@ -1,233 +1,0 @@
-"""
-Running NiBetaSeries using ds000164 (Stroop Task)
-===============================================================
-
-This example runs through a basic call of NiBetaSeries using
-the commandline entry point ``nibs``.
-While this example is using python, typically ``nibs`` will be
-called directly on the commandline.
-"""
-
-#############################################################################
-# Import all the necessary packages
-# =================================
-
-import tempfile  # make a temporary directory for files
-import os  # interact with the filesystem
-import urllib.request  # grad data from internet
-import tarfile  # extract files from tar
-from subprocess import Popen, PIPE, STDOUT  # enable calling commandline
-
-import matplotlib.pyplot as plt  # manipulate figures
-import seaborn as sns  # display results
-import pandas as pd   # manipulate tabular data
-
-#############################################################################
-# Download relevant data from ds000164 (and Atlas Files)
-# ======================================================
-# The subject data came from `openneuro <https://openneuro.org/datasets/ds000164/versions/00001/>`_.
-# The atlas data came from a `recently published parcellation <https://www.ncbi.nlm.nih.gov/pubmed/28981612>`_
-# in a publically accessible github repository.
-
-# atlas github repo for reference:
-"""https://github.com/ThomasYeoLab/CBIG/raw/master/stable_projects/\
-brain_parcellation/Schaefer2018_LocalGlobal/Parcellations/MNI/"""
-data_dir = tempfile.mkdtemp()
-print('Our working directory: {}'.format(data_dir))
-
-# download the tar data
-url = "https://www.dropbox.com/s/qoqbiya1ou7vi78/ds000164-test_v1.tar.gz?dl=1"
-tar_file = os.path.join(data_dir, "ds000164.tar.gz")
-u = urllib.request.urlopen(url)
-data = u.read()
-u.close()
-
-# write tar data to file
-with open(tar_file, "wb") as f:
-    f.write(data)
-
-# extract the data
-tar = tarfile.open(tar_file, mode='r|gz')
-tar.extractall(path=data_dir)
-
-os.remove(tar_file)
-
-#############################################################################
-# Display the minimal dataset necessary to run nibs
-# =================================================
-
-
-# https://stackoverflow.com/questions/9727673/list-directory-tree-structure-in-python
-def list_files(startpath):
-    for root, dirs, files in os.walk(startpath):
-        level = root.replace(startpath, '').count(os.sep)
-        indent = ' ' * 4 * (level)
-        print('{}{}/'.format(indent, os.path.basename(root)))
-        subindent = ' ' * 4 * (level + 1)
-        for f in files:
-            print('{}{}'.format(subindent, f))
-
-
-list_files(data_dir)
-
-#############################################################################
-# Manipulate events file so it satifies assumptions
-# =================================================
-# 1. the correct column has 1's and 0's corresponding to correct and incorrect,
-# respectively.
-# 2. the condition column is renamed to trial_type
-# nibs currently depends on the "correct" column being binary
-# and the "trial_type" column to contain the trial types of interest.
-
-#############################################################################
-# read the file
-# -------------
-
-events_file = os.path.join(data_dir,
-                           "ds000164",
-                           "sub-001",
-                           "func",
-                           "sub-001_task-stroop_events.tsv")
-events_df = pd.read_csv(events_file, sep='\t', na_values="n/a")
-print(events_df.head())
-
-#############################################################################
-# change the Y/N to 1/0
-# ---------------------
-
-events_df['correct'].replace({"Y": 1, "N": 0}, inplace=True)
-print(events_df.head())
-
[…]
