### 2019-04-17 32d8282
Tiny tweak to README
Add hyphens to "BIDS-compatible"

Yes I am a ridiculous pedant
