### 2019-04-23 43471be
minor clarifying statements
