### 2019-04-23 4e48e49
restructure docs for clarity
