+++ docs/betaseries.rst
@@ -4,30 +4,141 @@ Betaseries
 
 Introduction
 ------------
-Betaseries track the trial-to-trial modelled hemodynamic fluctuations
+Betaseries track the event-to-event modelled hemodynamic fluctuations
 that occur within task functional magnetic resonance imaging (task fMRI).
-This fills an important analytical gap between measuring hemodynamic
+Betaseries fills an important analytical gap between measuring hemodynamic
 fluctuations in resting state fMRI and measuring regional activations
 via cognitive subtraction in task fMRI.
 
+Conceptual Background
+---------------------
+Jesse Rissman [Rissman2004]_ was the first to publish on betaseries
+correlations describing their usage in a working memory task.
+In this task, participants saw a cue, a delay, and a probe, all occurring
+within a short time period.
+A cue was presented for one second, a delay occurred for seven seconds,
+and a probe was presented for one second.
+Given the Blood Oxygen-Level Dependent (BOLD) response
+takes approximately six seconds to reach its peak, and generally takes over
+20 seconds to return to baseline, we can begin to see a problem.
+The events within the events occur too close to each other to discern what
+brain responses are related to encoding the cue, the delay, or the probe.
+To discern how the activated brain regions form networks, Rissman
+created betaseries correlations.
+Instead of having a single regressor to describe all the cue events,
+a single regressor for all the delay events, and a single regressor for all the
+probe events (as is done in traditional task analysis),
+there is an individual regressor for every event in the experiment.
+For example, if your experiment has 40 events, each with a cue, delay, and
+probe event, the model will have a total of 120 regressors, fitting a beta
+(a.k.a. parameter) estimate for each event.
+Once you calculate a beta estimate for each event of a given event type
+(e.g. cue), at the end you will have four dimensional dataset where each volume
+represents the beta estimates for a particular event, and each voxel represents
+a specific beta estimate.
+
+Having one regressor per event in a single model is known as least squares all.
+This method, however, has limitations in the context of fast event related
+designs (e.g. designs where the events occur between 3-6
+seconds apart on average).
+Since each event has its own regressor, events that occur very close in time
+are colinear (e.g. are very overlapping).
+Jeanette Mumford et al. [Mumford2012]_ derived a solution for
+the high colinearity observed in least sqaures all by using another
+type of regression known as least squares separate.
+Instead of having one General Linear Model (GLM) with a regressor per event,
+least squares separate implements a GLM per event with only two regressors:
+1) one for the event of interest, and 2) one for every other event in the
+experiment.
+This process reduces the colinearity of the regressors and creates a more valid
+estimate of how each regressor fits the data.
+NiBetaSeries uses least squares separate and is thus optimized
+for fast event related designs.
+
+
+Math Background
+---------------
+.. math::
+   \begin{equation}
+        Y = X\beta + \epsilon
+    \end{equation}
+
+The above equation is the General Linear Model (GLM) presented using
+matrix notation.
+:math:`Y` represents the time-series we are attempting to explain.
+The :math:`\beta` assumes any value that minimizes the squared error between
+the modeled data and the actual data, :math:`Y`.
+Finally, :math:`\epsilon` (epsilon) refers to the error that is not captured
+by the model.
+Within a GLM, event-to-event betas may be averaged for a given event type
+and the variance is treated as noise.
+However, those event-to-event fluctuations may also contain important
+information the typical GLM will ignore/penalize.
+With a couple modifications to the above equation, we arrive at calculating a
+betaseries.
+
+.. math::
+    \begin{equation}
+        Y = X_1\beta_1 + X_2\beta_2 + . . . + X_n\beta_n + \epsilon
+    \end{equation}
+
+With the betaseries equation, a beta is estimated for every event, instead of
+for each event type (or whatever logical grouping).
[…]
