### 2019-04-24 2716066
incorrect pluralizations
I changed some pluralizations
