### 2019-04-27 51baaf7
new version docker & singularity
- `generate_nibetaseries_images.sh` was changed based on the suggestions from @jdkent [here]()
- as requested only `jupyter` and `seaborn` are installed via `conda` while the remaining python dependencies are gathered using `requirements.txt`  
- `nibetaseries` itself now follows `fmriprep`, that is the repository is copied into the `docker container`and subsequently installed (as in the [`fmriprep` Dockerfile]())
- this procedure needed the Dockerfile to be placed in the main `nibetaseries` location, hence the `docker` folder was deleted
