+++ Dockerfile
@@ -1,5 +1,5 @@
-# Generated by Neurodocker version 0.4.3-15-gad62560
-# Timestamp: 2019-04-15 12:05:01 UTC
+# Generated by Neurodocker version 0.4.3-16-g21e55b3
+# Timestamp: 2019-04-27 17:39:59 UTC
 # 
 # Thank you for using Neurodocker. If you discover any issues
 # or ways to improve this software, please submit an issue or
@@ -24,7 +24,7 @@ RUN export ND_ENTRYPOINT="/neurodocker/startup.sh" \
            locales \
            unzip \
     && apt-get clean \
-    && rm -rf /var/lib/apt/lists/* /tmp/* /var/tmp/* \
+    && rm -rf /var/lib/apt/lists/* \
     && sed -i -e 's/# en_US.UTF-8 UTF-8/en_US.UTF-8 UTF-8/' /etc/locale.gen \
     && dpkg-reconfigure --frontend=noninteractive locales \
     && update-locale LANG="en_US.UTF-8" \
@@ -45,7 +45,7 @@ RUN apt-get update -qq \
            gcc \
            g++ \
     && apt-get clean \
-    && rm -rf /var/lib/apt/lists/* /tmp/* /var/tmp/*
+    && rm -rf /var/lib/apt/lists/*
 
 RUN test "$(getent passwd neuro)" || useradd --no-user-group --create-home --shell /bin/bash neuro
 USER neuro
@@ -69,24 +69,31 @@ RUN export PATH="/opt/miniconda-latest/bin:$PATH" \
     && sync && conda clean --all && sync \
     && bash -c "source activate neuro_py36 \
     &&   pip install --no-cache-dir  \
-             'ipython' \
-             'nipype==1.1.5' \
-             'pybids==0.6.4' \
-             'nibabel==2.3.0' \
-             'nistats==0.0.1b' \
-             'nilearn==0.4.2' \
-             'pandas==0.24.0' \
-             'numpy==1.14.5' \
-             'duecredit==0.6.4' \
-             'scikit-learn==0.19.2' \
              'jupyter' \
-             'seaborn' \
-             'nibetaseries'" \
+             'seaborn'" \
     && rm -rf ~/.cache/pip/* \
     && sync \
     && sed -i '$isource activate neuro_py36' $ND_ENTRYPOINT
 
+COPY ["requirements.txt", "/src/"]
+
+COPY [".", "/src/nibetaseries"]
+
+WORKDIR /src/
+
+USER root
+
+RUN chmod -R 777 requirements.txt && chmod -R 777 /src/nibetaseries/
+
+USER neuro
+
+RUN pip install --user --no-cache-dir -r requirements.txt
+
+RUN cd /src/nibetaseries && pip install --no-cache-dir .[all]
+
 RUN mkdir -p ~/.jupyter && echo c.NotebookApp.ip = \"0.0.0.0\" > ~/.jupyter/jupyter_notebook_config.py
+
+WORKDIR /home
 
 RUN echo '{ \
     \n  "pkg_manager": "apt", \
@@ -113,27 +120,58 @@ RUN echo '{ \
     \n          "python=3.6" \
     \n        ], \
     \n        "pip_install": [ \
-    \n          "ipython", \
-    \n          "nipype==1.1.5", \
-    \n          "pybids==0.6.4", \
-    \n          "nibabel==2.3.0", \
-    \n          "nistats==0.0.1b", \
-    \n          "nilearn==0.4.2", \
-    \n          "pandas==0.24.0", \
-    \n          "numpy==1.14.5", \
-    \n          "duecredit==0.6.4", \
-    \n          "scikit-learn==0.19.2", \
     \n          "jupyter", \
[…]
