### 2019-04-27 a0eae25
uncomment singularity related commands
- the singularity related commands `generate_singularity > Singularity` & `singularity build nibetaseries.simg Singularity` within `generate_nibetaseries_images.sh` were uncommented
