### 2019-04-29 c175649
Suggestions for improved clarity
