### 2019-05-02 a9a6adb
changes to MANIFEST.in & update Singularity
- the following lines were added to MANIFEST.in in order to fix travisci error: exclude *.sh, exclude Dockerfile, exclude Singularity
- the Singularity file wasn't updated in the last commit, that is regenerated by generate_nibetaseries_images.sh, which was done within the commit at hand
