### 2019-05-10 8bee9af
remove singularity build and modify docker
- simplify the install
- assume we can build singularity images from the docker image
