+++ Dockerfile
@@ -1,5 +1,5 @@
-# Generated by Neurodocker version 0.4.3-16-g21e55b3
-# Timestamp: 2019-05-02 12:31:45 UTC
+# Generated by Neurodocker version 0.5.0
+# Timestamp: 2019-05-10 19:08:05 UTC
 # 
 # Thank you for using Neurodocker. If you discover any issues
 # or ways to improve this software, please submit an issue or
@@ -44,6 +44,7 @@ RUN apt-get update -qq \
     && apt-get install -y -q --no-install-recommends \
            gcc \
            g++ \
+           graphviz \
     && apt-get clean \
     && rm -rf /var/lib/apt/lists/*
 
@@ -70,30 +71,27 @@ RUN export PATH="/opt/miniconda-latest/bin:$PATH" \
     && bash -c "source activate neuro_py36 \
     &&   pip install --no-cache-dir  \
              'jupyter' \
-             'seaborn'" \
+             'seaborn' \
+             'tox'" \
     && rm -rf ~/.cache/pip/* \
     && sync \
     && sed -i '$isource activate neuro_py36' $ND_ENTRYPOINT
 
-COPY ["requirements.txt", "/src/"]
-
 COPY [".", "/src/nibetaseries"]
-
-WORKDIR /src/
 
 USER root
 
-RUN chmod -R 777 requirements.txt && chmod -R 777 /src/nibetaseries/
+RUN chown -R neuro:users /src/nibetaseries
 
 USER neuro
 
-RUN pip install --user --no-cache-dir -r requirements.txt
+WORKDIR /src/nibetaseries
 
-RUN cd /src/nibetaseries && pip install --no-cache-dir .[all]
+RUN pip install --no-cache-dir .
 
 RUN mkdir -p ~/.jupyter && echo c.NotebookApp.ip = \"0.0.0.0\" > ~/.jupyter/jupyter_notebook_config.py
 
-WORKDIR /home
+WORKDIR /home/neuro
 
 RUN echo '{ \
     \n  "pkg_manager": "apt", \
@@ -106,7 +104,8 @@ RUN echo '{ \
     \n      "install", \
     \n      [ \
     \n        "gcc", \
-    \n        "g++" \
+    \n        "g++", \
+    \n        "graphviz" \
     \n      ] \
     \n    ], \
     \n    [ \
@@ -121,18 +120,12 @@ RUN echo '{ \
     \n        ], \
     \n        "pip_install": [ \
     \n          "jupyter", \
-    \n          "seaborn" \
+    \n          "seaborn", \
+    \n          "tox" \
     \n        ], \
     \n        "create_env": "neuro_py36", \
     \n        "activate": true \
     \n      } \
-    \n    ], \
-    \n    [ \
-    \n      "copy", \
-    \n      [ \
-    \n        "requirements.txt", \
-    \n        "/src/" \
-    \n      ] \
     \n    ], \
     \n    [ \
     \n      "copy", \
@@ -142,28 +135,24 @@ RUN echo '{ \
     \n      ] \
     \n    ], \
     \n    [ \
-    \n      "workdir", \
-    \n      "/src/" \
-    \n    ], \
-    \n    [ \
     \n      "user", \
     \n      "root" \
     \n    ], \
     \n    [ \
     \n      "run", \
-    \n      "chmod -R 777 requirements.txt && chmod -R 777 /src/nibetaseries/" \
+    \n      "chown -R neuro:users /src/nibetaseries" \
     \n    ], \
     \n    [ \
     \n      "user", \
     \n      "neuro" \
     \n    ], \
     \n    [ \
-    \n      "run", \
-    \n      "pip install --user --no-cache-dir -r requirements.txt" \
[…]
