### 2019-05-10 6d4d6d7
remove singularity and add circleci
