### 2019-05-10 a013815
unindent docker_tox steps
