### 2019-05-10 5620e1d
add nibs to test command
