+++ .circleci/config.yml
@@ -46,7 +46,7 @@ jobs:
       - run:
           name: Docker save
           no_output_timeout: 30m
-          command:
+          command: |
             mkdir -p /tmp/cache
             docker save hbclab/nibetaseries:latest | pigz -3 > /tmp/cache/docker.tar.gz
       
