+++ .circleci/config.yml
@@ -70,15 +70,15 @@ jobs:
           keys:
             - docker-v1-{{ .Branch }}-{{ .Revision }}
       - run:
-        name: Load Docker image layer cache
-        no_output_timeout: 30m
-        command: |
-          docker info
-          set +o pipefail
-          if [ -f /tmp/cache/docker.tar.gz ]; then
-            sudo apt update && sudo apt -y install pigz
-            pigz -d --stdout /tmp/cache/docker.tar.gz | docker load
-            docker images
+          name: Load Docker image layer cache
+          no_output_timeout: 30m
+          command: |
+            docker info
+            set +o pipefail
+            if [ -f /tmp/cache/docker.tar.gz ]; then
+              sudo apt update && sudo apt -y install pigz
+              pigz -d --stdout /tmp/cache/docker.tar.gz | docker load
+              docker images
           fi
       - run:
           name: Run NiBetaSeries tests
