+++ .circleci/config.yml
@@ -79,7 +79,7 @@ jobs:
               sudo apt update && sudo apt -y install pigz
               pigz -d --stdout /tmp/cache/docker.tar.gz | docker load
               docker images
-          fi
+            fi
       - run:
           name: Run NiBetaSeries tests
           command: |
