### 2019-05-21 #1295 @jdkent
Hi @danielskatz, I am working through the comments still. I should have more time to dedicate to the project this week and the next.
