### 2019-05-23 fc6a2a9
update install instructions
- the install instructions in the README.rst and the respective part of the docs, that is docs/installation.rst were updated to include information regarding the containerized versions (Docker and Singularity)
- additionally, example calls of both the containerized versions and pure python were added
