### 2019-06-09 ee03341
only allow pip to install if above 3.5
