+++ setup.py
@@ -70,6 +70,7 @@ setup(
         #   'rst': ['docutils>=0.11'],
         #   ':python_version=="2.6"': ['argparse'],
     },
+    python_requires='~=3.5',
     entry_points={
         'console_scripts': [
             'nibs = nibetaseries.cli.run:main',
