### 2019-06-09 05c1e2f
add headers to docs
- also edit text in installation
