+++ docs/workflows.rst
@@ -1,3 +1,5 @@
+.. _workflows:
+
 =========
 Workflows
 =========
