### 2019-06-09 a855b11
[ENH] Add Docker and Singularity Support (#140)
* fix pybids link

* initial draft of docker / singularity image

* new version docker & singularity

* remove singularity build and modify docker

* remove sphinx versioning dependency

* add circleci file

* trigger circleci

* update install instructions

* remove rst check in check test

* add auto_example to .gitignore
