+++ docs/index.rst
@@ -1,3 +1,5 @@
+.. _index:
+
 ========
 Contents
 ========
