+++ CHANGELOG.rst
@@ -1,3 +1,5 @@
+.. _changelog:
+
 0.2.3 (January 29, 2019)
 ========================
 
