+++ AUTHORS.rst
@@ -1,3 +1,4 @@
+.. _authors:
 
 Authors
 =======
