### 2019-06-12 fa93e6f
add output for the figure
