### 2019-06-12 65ade25
create datasink interface for figure output and connect to workflow
