+++ src/nibetaseries/workflows/base.py
@@ -196,6 +196,8 @@ def init_single_subject_wf(atlas_img, atlas_lut, bold_metadata_list, brainmask_l
 
         correlation_matrix
             a matrix (tsv) file denoting all roi-roi correlations
+        correlation_fig
+            a jpg file of a circular connectivity plot showing all roi-roi correlations
     """
     workflow = pe.Workflow(name=name)
 
@@ -235,6 +237,11 @@ def init_single_subject_wf(atlas_img, atlas_lut, bold_metadata_list, brainmask_l
                                        iterfield=['betaseries_file', 'in_file'],
                                        name='ds_correlation_matrix')
 
+    ds_correlation_fig = pe.MapNode(DerivativesDataSink(base_directory=output_dir,
+                                                        suffix='fig'),
+                                    iterfield=['betaseries_file', 'in_file'],
+                                    name='ds_correlation_fig')
+
     # connect the nodes
     workflow.connect([
         (input_node, betaseries_wf,
@@ -247,10 +254,15 @@ def init_single_subject_wf(atlas_img, atlas_lut, bold_metadata_list, brainmask_l
             [('output_node.betaseries_files', 'input_node.betaseries_files')]),
         (input_node, correlation_wf, [('atlas_img', 'input_node.atlas_file'),
                                       ('atlas_lut', 'input_node.atlas_lut')]),
+
         (correlation_wf, output_node, [('output_node.correlation_matrix', 'correlation_matrices')]),
         (input_node, ds_correlation_matrix, [('preproc_img', 'source_file')]),
         (betaseries_wf, ds_correlation_matrix, [('output_node.betaseries_files', 'betaseries_file')]),
         (output_node, ds_correlation_matrix, [('correlation_matrices', 'in_file')]),
+
+        (input_node, ds_correlation_fig, [('preproc_img', 'source_file')]),
+        (betaseries_wf, ds_correlation_fig, [('output_node.betaseries_files', 'betaseries_file')]),
+        (output_node, ds_correlation_fig, [('correlation_fig', 'in_file')])
     ])
 
     return workflow
