### 2019-06-12 d527789
make readme the landing page for docs
