+++ docs/index.rst
@@ -1,27 +1,1 @@
-.. _index:
-
-========
-Contents
-========
-
-.. toctree::
-   :maxdepth: 2
-
-   readme
-   betaseries
-   installation
-   usage
-   auto_examples/index
-   workflows
-   reference/index
-   contributing
-   authors
-   changelog
-
-Indices and tables
-==================
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
-
+.. include:: ../README.rst
