### 2019-06-13 e40b9e7
tweak figure generation - change png to svg - do mot make figure interactive
