+++ src/nibetaseries/interfaces/nilearn.py
@@ -21,7 +21,7 @@ class AtlasConnectivityInputSpec(BaseInterfaceInputSpec):
 
 class AtlasConnectivityOutputSpec(TraitedSpec):
     correlation_matrix = File(exists=True, desc='roi-roi fisher z transformed correlation matrix')
-    correlation_fig = File(exists=True, desc='jpg of roi-roi fisher z transformed correlation matrix')
+    correlation_fig = File(exists=True, desc='svg of roi-roi fisher z transformed correlation matrix')
 
 
 class AtlasConnectivity(NilearnBaseInterface, SimpleInterface):
@@ -73,14 +73,14 @@ class AtlasConnectivity(NilearnBaseInterface, SimpleInterface):
         # define title and outfile names:
         trial_regex = re.compile(r'.*trialtype-(?P<trial>[A-Za-z0-9]+)')
         title = re.search(trial_regex, self.inputs.timeseries_file).groupdict()['trial']
-        outfile = os.path.join(runtime.cwd, ".".join([trial_regex, "jpg"]))
+        outfile = os.path.join(runtime.cwd, ".".join([title, "svg"]))
 
         n_lines = int(np.sum(connmat > 0) / 2)
         fig = plt.figure(figsize=(5, 5))
 
         plot_connectivity_circle(connmat, labels, n_lines=n_lines, fig=fig, title=title, fontsize_title=10,
                                  facecolor='white', textcolor='black', colormap='jet', colorbar=1,
-                                 node_colors=['black'], node_edgecolor=['white'])
+                                 node_colors=['black'], node_edgecolor=['white'], show=False)
 
         fig.savefig(outfile, dpi=300)
         self._results['correlation_fig'] = outfile
