### 2019-06-13 df5372b
fix naming and connections
