+++ src/nibetaseries/workflows/analysis.py
@@ -49,22 +49,9 @@ def init_correlation_wf(name="correlation_wf"):
         correlation_matrix
             a matrix (tsv) file denoting all roi-roi correlations
         correlation_fig
-            a jpg file of a circular connectivity plot showing all roi-roi correlations
+            a svg file of a circular connectivity plot showing all roi-roi correlations
     """
     workflow = pe.Workflow(name=name)
-
-    def _rename_matrix(correlation_matrix, betaseries_file):
-        import os
-        import re
-        from shutil import copyfile
-        # import pdb; pdb.set_trace()
-        betaseries_regex = re.compile('.*betaseries_trialtype-(?P<trial_type>[A-Za-z0-9_]+).nii.gz')
-        trial_type = betaseries_regex.search(betaseries_file).groupdict()['trial_type'].replace('_', '')
-        out_file = os.path.join(os.getcwd(),
-                                'correlation-matrix_trialtype-{trial_type}.tsv'.format(trial_type=trial_type))
-        copyfile(correlation_matrix, out_file)
-
-        return out_file
 
     input_node = pe.MapNode(niu.IdentityInterface(fields=['betaseries_files',
                                                           'atlas_file',
@@ -93,3 +80,17 @@ def init_correlation_wf(name="correlation_wf"):
     ])
 
     return workflow
+
+
+def _rename_matrix(correlation_matrix, betaseries_file):
+    import os
+    import re
+    from shutil import copyfile
+    # import pdb; pdb.set_trace()
+    betaseries_regex = re.compile('.*betaseries_trialtype-(?P<trial_type>[A-Za-z0-9_]+).nii.gz')
+    trial_type = betaseries_regex.search(betaseries_file).groupdict()['trial_type'].replace('_', '')
+    out_file = os.path.join(os.getcwd(),
+                            'correlation-matrix_trialtype-{trial_type}.tsv'.format(trial_type=trial_type))
+    copyfile(correlation_matrix, out_file)
+
+    return out_file
