### 2019-06-13 e892a0b
fix naming and connections
