### 2019-06-13 74168a9
fix tkinter error msg: tkinter.TclError: no display name and no $DISPLAY environment variable
