### 2019-06-13 37f206b
Revert "fix tkinter error"
This reverts commit 74168a92f0a1af3399dec188b51347dc7ae1da67.
