### 2019-06-13 73eb093
set dummy display variable
