### 2019-06-13 06471bc
make image not interactive
