### 2019-06-13 58e0438
remove display variable
