+++ .travis.yml
@@ -6,7 +6,6 @@ env:
   global:
   - LD_PRELOAD=/lib/x86_64-linux-gnu/libSegFault.so
   - SEGFAULT_SIGNALS=all
-  - DISPLAY=:99.0
   - secure: "VovtWoTavKAeC15okVReKwL+c7Lbx/I0nJhFfMoTqO9DtXSimg2IjWLEv6XyD57/3a277NwUAOIq6j1qEnPRhjnGeUwY5bU1FF9txk0F5uLsc8DC0brzaMrPuOxgyc9nyVjTtkaKD4PWfH/8BlqJNdNoe/onsktmETfEfF4rYVCbsHCOLxi2XLOK5WdTvd9OJvjyT7doNuDhMinWPWw73Ah4qhAcqJULexAgWEypSwtssBXOFCTdhkHgRj0z7HZkUpfQILp3yh6rJS0rCo7VMVj7f+1IgTPcvyx/6cez88hUZYu6n0gBTYLGyIhIYWXt0nxpXL8FEWonNSiiJkKt7cuCkLt857hkF0xI5McHI0ETvqc97ICzUjae496nnuP6NZgTGEzcqHUS46KQh3faGzKA8ptkEllCn9683usZ5HF6a1GMAr/NXdmY/pGOhZO2zGMLt2Y8S9/XPLIe39HlTkoAMjBYb6lOQ5w/v3uA0Q3awuMzM0Dqm+BldB+9GTijWzKlre04dfJeIja3RZrGI83yUaMp7/NasKGVYulIT4C/3epPZIR/UmUDGul6xwQmpzplzVYJVNRr0s2BTbnWVjxc+h/doofx3UWTlItti3mH9B0Xq6klTRR15cjY/mh304OsUvwNwS7Ic5QtdMOkhTlGu/uegjrU0QCd6eNUemI="
 matrix:
   include:
