### 2019-07-02 8977846
[ENH] add visualizations (#148)
* add visualizations

* fix flake8 errors

* Update nilearn.py

* fix Travis CI errors

* add output for the figure

* create datasink interface for figure output
and connect to workflow

* add mne and matplotlib

* tweak figure generation
- change png to svg
- do mot make figure interactive

* fix naming and connections

* use agg backend
