+++ src/nibetaseries/workflows/analysis.py
@@ -48,21 +48,10 @@ def init_correlation_wf(name="correlation_wf"):
 
         correlation_matrix
             a matrix (tsv) file denoting all roi-roi correlations
+        correlation_fig
+            a svg file of a circular connectivity plot showing all roi-roi correlations
     """
     workflow = pe.Workflow(name=name)
-
-    def _rename_matrix(correlation_matrix, betaseries_file):
-        import os
-        import re
-        from shutil import copyfile
-        # import pdb; pdb.set_trace()
-        betaseries_regex = re.compile('.*betaseries_trialtype-(?P<trial_type>[A-Za-z0-9_]+).nii.gz')
-        trial_type = betaseries_regex.search(betaseries_file).groupdict()['trial_type'].replace('_', '')
-        out_file = os.path.join(os.getcwd(),
-                                'correlation-matrix_trialtype-{trial_type}.tsv'.format(trial_type=trial_type))
-        copyfile(correlation_matrix, out_file)
-
-        return out_file
 
     input_node = pe.MapNode(niu.IdentityInterface(fields=['betaseries_files',
                                                           'atlas_file',
@@ -70,7 +59,8 @@ def init_correlation_wf(name="correlation_wf"):
                             iterfield=['betaseries_files'],
                             name='input_node')
 
-    output_node = pe.Node(niu.IdentityInterface(fields=['correlation_matrix']),
+    output_node = pe.Node(niu.IdentityInterface(fields=['correlation_matrix',
+                                                        'correlation_fig']),
                           name='output_node')
 
     atlas_corr_node = pe.MapNode(AtlasConnectivity(), name='atlas_corr_node', iterfield=['timeseries_file'])
@@ -86,6 +76,21 @@ def init_correlation_wf(name="correlation_wf"):
         (input_node, rename_matrix_node, [('betaseries_files', 'betaseries_file')]),
         (atlas_corr_node, rename_matrix_node, [('correlation_matrix', 'correlation_matrix')]),
         (rename_matrix_node, output_node, [('correlation_matrix_trialtype', 'correlation_matrix')]),
+        (atlas_corr_node, output_node, [('correlation_fig', 'correlation_fig')])
     ])
 
     return workflow
+
+
+def _rename_matrix(correlation_matrix, betaseries_file):
+    import os
+    import re
+    from shutil import copyfile
+    # import pdb; pdb.set_trace()
+    betaseries_regex = re.compile('.*betaseries_trialtype-(?P<trial_type>[A-Za-z0-9_]+).nii.gz')
+    trial_type = betaseries_regex.search(betaseries_file).groupdict()['trial_type'].replace('_', '')
+    out_file = os.path.join(os.getcwd(),
+                            'correlation-matrix_trialtype-{trial_type}.tsv'.format(trial_type=trial_type))
+    copyfile(correlation_matrix, out_file)
+
+    return out_file
