### 2019-07-02 07c6a67
require python3 (#147)
* only allow pip to install if above 3.5

* change reference to index to be contents

* make readme the landing page for docs
