+++ docs/contents.rst
@@ -1,0 +1,28 @@
+:orphan:
+
+.. _contents:
+
+========
+Contents
+========
+
+.. toctree::
+   :maxdepth: 2
+
+   betaseries
+   installation
+   usage
+   auto_examples/index
+   workflows
+   reference/index
+   contributing
+   authors
+   changelog
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
