### 2019-07-02 985abb8
temporary fix to dockerfile
