### 2019-07-02 74808d5
add build_docs to workflow
