### 2019-07-02 1ee6ae7
fix dockerfile and doc build
