+++ Dockerfile
@@ -1,5 +1,5 @@
 # Generated by Neurodocker version 0.5.0
-# Timestamp: 2019-05-10 19:08:05 UTC
+# Timestamp: 2019-07-02 21:26:54 UTC
 # 
 # Thank you for using Neurodocker. If you discover any issues
 # or ways to improve this software, please submit an issue or
@@ -51,6 +51,8 @@ RUN apt-get update -qq \
 RUN test "$(getent passwd neuro)" || useradd --no-user-group --create-home --shell /bin/bash neuro
 USER neuro
 
+WORKDIR /home/neuro
+
 ENV CONDA_DIR="/opt/miniconda-latest" \
     PATH="/opt/miniconda-latest/bin:$PATH"
 RUN export PATH="/opt/miniconda-latest/bin:$PATH" \
@@ -65,7 +67,8 @@ RUN export PATH="/opt/miniconda-latest/bin:$PATH" \
     && conda config --system --set show_channel_urls true \
     && sync && conda clean --all && sync \
     && conda create -y -q --name neuro_py36 \
-    && (conda install -y -q --name neuro_py36 'python=3.6' || true) \
+    && conda install -y -q --name neuro_py36 \
+           'python=3.6' \
     && sync && conda clean --all && sync \
     && bash -c "source activate neuro_py36 \
     &&   pip install --no-cache-dir  \
@@ -110,6 +113,10 @@ RUN echo '{ \
     \n    [ \
     \n      "user", \
     \n      "neuro" \
+    \n    ], \
+    \n    [ \
+    \n      "workdir", \
+    \n      "/home/neuro" \
     \n    ], \
     \n    [ \
     \n      "miniconda", \
