### 2019-07-02 ffe12e2
update/add references in joss paper
