+++ .joss/paper.bib
@@ -74,3 +74,37 @@ url = {https://github.com/poldrack/pybetaseries},
 title = {{pybetaseries}},
 year = {2014}
 }
+
+@article{Abdulrahman2016,
+abstract = {Functional magnetic resonance imaging (fMRI) studies typically employ rapid, event-related designs for behavioral reasons and for reasons associated with statistical efficiency. Efficiency is calculated from the precision of the parameters (Betas) estimated from a General Linear Model (GLM) in which trial onsets are convolved with a Hemodynamic Response Function (HRF). However, previous calculations of efficiency have ignored likely variability in the neural response from trial to trial, for example due to attentional fluctuations, or different stimuli across trials. Here we compare three GLMs in their efficiency for estimating average and individual Betas across trials as a function of trial variability, scan noise and Stimulus Onset Asynchrony (SOA): "Least Squares All" (LSA), "Least Squares Separate" (LSS) and "Least Squares Unitary" (LSU). Estimation of responses to individual trials in particular is important for both functional connectivity using "Beta-series correlation" and "multi-voxel pattern analysis" (MVPA). Our simulations show that the ratio of trial-to-trial variability to scan noise impacts both the optimal SOA and optimal GLM, especially for short SOAs. {\textless}. 5. s: LSA is better when this ratio is high, whereas LSS and LSU are better when the ratio is low. For MVPA, the consistency across voxels of trial variability and of scan noise is also critical. These findings not only have important implications for design of experiments using Beta-series regression and MVPA, but also statistical parametric mapping studies that seek only efficient estimation of the mean response across trials.},
+author = {Abdulrahman, Hunar and Henson, Richard N.},
+doi = {10.1016/j.neuroimage.2015.11.009},
+file = {:media/Data/Downloads/1-s2.0-S1053811915010319-main.pdf:pdf},
+issn = {10959572},
+journal = {NeuroImage},
+keywords = {Bold variability,FMRI design,General Linear Model,Least squares all,Least squares separate,MVPA,Trial based correlations},
+mendeley-groups = {Prospectus},
+pages = {756--766},
+publisher = {The Authors},
+title = {{Effect of trial-to-trial variability on optimal event-related fMRI design: Implications for Beta-series correlation and multi-voxel pattern analysis}},
+url = {http://dx.doi.org/10.1016/j.neuroimage.2015.11.009},
+volume = {125},
+year = {2016}
+}
+
+@article{Turner2012a,
[…]
