### 2019-07-02 2e64712
add references and use bibliography
