+++ docs/references.bib
@@ -1,0 +1,127 @@
+@article{Gottlich2015,
+abstract = {BASCO (BetA Series COrrelation) is a user-friendly MATLAB toolbox with a graphical user interface (GUI) which allows investigating functional connectivity in event-related functional magnetic resonance imaging (fMRI) data. Connectivity analyses extend and compliment univariate activation analyses since the actual interaction between brain regions involved in a task can be explored. BASCO supports seed-based functional connectivity as well as brain network analyses. Although there are a multitude of advanced toolboxes for investigating resting-state functional connectivity, BASCO is the first toolbox for evaluating task-related whole-brain functional connectivity employing a large number of network nodes. Thus, BASCO allows investigating task-specific rather than resting-state networks. Here, we summarize the main features of the toolbox and describe the methods and algorithms.},
+author = {G{\"{o}}ttlich, Martin and Beyer, Frederike and Kr{\"{a}}mer, Ulrike M.},
+doi = {10.3389/fnsys.2015.00126},
+isbn = {1662-5137 (Electronic)$\backslash$r1662-5137 (Linking)},
+issn = {1662-5137},
+journal = {Frontiers in Systems Neuroscience},
+keywords = {b,beta-series correlation,betaseries,degree centrality,event-related design,fMRI,fmri,functional connectivity,method,network analysis},
+mendeley-tags = {betaseries,method},
+number = {September},
+pages = {1--10},
+pmid = {26441558},
+title = {{BASCO: a toolbox for task-related functional connectivity}},
+url = {http://journal.frontiersin.org/Article/10.3389/fnsys.2015.00126/abstract},
+volume = {9},
+year = {2015}
+}
+@article{Esteban2018a,
+abstract = {Preprocessing of functional MRI (fMRI) involves numerous steps to clean and standardize data before statistical analysis. Generally, researchers create ad hoc preprocessing workflows for each new dataset, building upon a large inventory of tools available for each step. The complexity of these workflows has snowballed with rapid advances in MR data acquisition and image processing techniques. We introduce fMRIPrep, an analysis-agnostic tool that addresses the challenge of robust and reproducible preprocessing for task-based and resting fMRI data. FMRIPrep automatically adapts a best-in-breed workflow to the idiosyncrasies of virtually any dataset, ensuring high-quality preprocessing with no manual intervention. By introducing visual assessment checkpoints into an iterative integration framework for software-testing, we show that fMRIPrep robustly produces high-quality results on a diverse fMRI data collection comprising participants from 54 different studies in the OpenfMRI repository. We review the distinctive features of fMRIPrep in a qualitative comparison to other preprocessing workflows. We demonstrate that fMRIPrep achieves higher spatial accuracy as it introduces less uncontrolled spatial smoothness than one commonly used preprocessing tool. FMRIPrep has the potential to transform fMRI research by equipping neuroscientists with a high-quality, robust, easy-to-use and transparent preprocessing workflow which can help ensure the validity of inference and the interpretability of their results.},
+archivePrefix = {arXiv},
+arxivId = {1201.0490},
[…]
