+++ docs/betaseries.rst
@@ -16,7 +16,7 @@ via cognitive subtraction in task fMRI.
 
 Conceptual Background
 ---------------------
-Jesse Rissman [Rissman2004]_ was the first to publish on betaseries
+Jesse Rissman :cite:`Rissman2004` was the first to publish on betaseries
 correlations describing their usage in a working memory task.
 In this task, participants saw a cue, a delay, and a probe, all occurring
 within a short time period.
@@ -47,7 +47,7 @@ designs (e.g. designs where the events occur between 3-6
 seconds apart on average).
 Since each event has its own regressor, events that occur very close in time
 are colinear (e.g. are very overlapping).
-Jeanette Mumford et al. [Mumford2012]_ derived a solution for
+Benjamin Turner :cite:`Turner2012a` derived a solution for
 the high colinearity observed in least sqaures all by using another
 type of regression known as least squares separate.
 Instead of having one General Linear Model (GLM) with a regressor per event,
@@ -201,24 +201,17 @@ Relevent Software
 BASCO_ (BetA Series COrrelations) is a matlab program that also performs
 betaseries correlations
 
+.. _BASCO: https://www.nitrc.org/projects/basco/
 
 Other Relevant Readings
 -----------------------
-- [Cisler2014]_
-- [Gottlich2015]_
-- [Abdulrahman2016]_
+- :cite:`Cisler2012`: A comparison of psychophysiological interactions and LSS
+- :cite:`Gottlich2015`: The BASCO paper
+- :cite:`Abdulrahman2016`: evaluation of LSS (and other methods)
+
 
 References
 ----------
 
-.. [Rissman2004] https://www.ncbi.nlm.nih.gov/pubmed/15488425
-
-.. [Mumford2012] https://www.ncbi.nlm.nih.gov/pubmed/21924359
-
-.. [Cisler2014] https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4019671/
-
-.. [Gottlich2015] https://www.frontiersin.org/articles/10.3389/fnsys.2015.00126/full
-
-.. [Abdulrahman2016] https://www.ncbi.nlm.nih.gov/pubmed/26549299
-
-.. _BASCO: https://www.nitrc.org/projects/basco/
+.. bibliography:: references.bib
+    :style: plain
