### 2019-07-02 d8a1d02
add pybetaseries and fix #117
