### 2019-07-02 cf376ea
fix bullet list syntax
