### 2019-07-03 212bcf6
update documentation - delete contents.rst - update references - make the bash command more clear in the tutorial
