+++ docs/references.bib
@@ -125,3 +125,16 @@ title = {{A Comparison of Statistical Methods for Detecting Context- Modulated F
 volume = {24},
 year = {2014}
 }
+@article{Verstynen2014,
+abstract = { Accurately making a decision in the face of incongruent options increases the efficiency of making similar congruency decisions in the future. Contextual factors like reward can modulate this adaptive process, suggesting that networks associated with monitoring previous success and failure outcomes might contribute to this form of behavioral updating. To evaluate this possibility, a group of healthy adults (n = 30) were tested with functional MRI (fMRI) while they performed a color-word Stroop task. In a conflict-related region of the medial orbitofrontal cortex (mOFC), stronger BOLD responses predicted faster response times (RTs) on the next trial. More importantly, the degree of behavioral adaptation of RTs was correlated with the magnitude of mOFC-RT associations on the previous trial, but only after accounting for network-level interactions with prefrontal and striatal regions. This suggests that congruency sequencing effects may rely on interactions between distributed corticostriatal circuits. This possibility was evaluated by measuring the convergence of white matter projections from frontal areas into the striatum with diffusion-weighted imaging. In these pathways, greater convergence of corticostriatal projections correlated with stronger functional mOFC-RT associations that, in turn, provided an indirect pathway linking anatomical structure to behavior. Thus distributed corticostriatal processing may mediate the orbitofrontal cortex's influence on behavioral updating, even in the absence of explicit rewards. },
+annote = {PMID: 25143543},
+author = {Verstynen, Timothy D},
+doi = {10.1152/jn.00221.2014},
+journal = {Journal of Neurophysiology},
+number = {10},
+pages = {2457--2469},
+title = {{The organization and dynamics of corticostriatal pathways link the medial orbitofrontal cortex to future behavioral responses}},
+url = {https://doi.org/10.1152/jn.00221.2014},
+volume = {112},
+year = {2014}
+}
