### 2019-07-03 e5bf839
add link to tutorial from readme.rst
