### 2019-07-03 1b3288e
add concrete example of command
