### 2019-07-03 7829479
update example tutorial with hyphen
