### 2019-07-03 17ec84d
[DOC] add references (#153)
* update/add references in joss paper

* add references and use bibliography

* add pybetaseries and fix #117

* fix bullet list syntax

* simplify docs build

* update documentation
- delete contents.rst
- update references
- make the bash command more clear in the tutorial

* add link to tutorial from readme.rst
