+++ docs/index.rst
@@ -1,1 +1,25 @@
 .. include:: ../README.rst
+
+Contents
+========
+
+.. toctree::
+   :maxdepth: 2
+
+   betaseries
+   installation
+   usage
+   auto_examples/index
+   workflows
+   reference/index
+   contributing
+   authors
+   changelog
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
