+++ docs/usage.rst
@@ -19,6 +19,24 @@ Example Call(s)
 If you want to see a full example on preparing data and running ``nibs``, see:
 :ref:`sphx_glr_auto_examples_plot_run_nibetaseries.py`.
 
+Example 1
+~~~~~~~~~
+
+.. code-block:: bash
+
+    nibs \
+    /home/james/bids/ \
+    fmriprep \
+    /home/james/bids/derivatives/betaSeries/schaefer_parcel-400_network-17 \
+    participant
+    -a /home/james/bids/derivatives/atlas/Schaefer2018_400Parcels_17Networks_order_FSLMNI152_2mm.nii.gz \
+    -l /home/james/bids/derivatives/atlas/schaefer_parcel-400_network-17.tsv \
+    -w /home/james/bids/derivatives/betaSeries/work_n17 \
+    -c WhiteMatter CSF Cosine01 Cosine02 Cosine03 Cosine04 Cosine05 Cosine06 Cosine07 \
+    --nthreads 32 \
+    --variant AROMAnonaggr \
+    --hrf-model 'glover + derivative + dispersion'
+
 In this example we have our bids directory ``/home/james/bids``,
 and our derivatives directory ``/home/james/bids/derivatives``.
 ``nibs`` currently assumes that ``derivatives`` is immediately underneath
@@ -55,21 +73,3 @@ the bids directory.
   by ``ICA-AROMA``.
 - the hrf model (``--hrf-model``) passes all the available options
   from `nistats <https://nistats.github.io/index.html>`_.
-
-Given all the above information, the actual ``nibs`` command would
-look like the following:
-
-.. code-block:: bash
-
-    nibs \
-    /home/james/bids/ \
-    fmriprep \
-    /home/james/bids/derivatives/betaSeries/schaefer_parcel-400_network-17 \
-    participant
-    -a /home/james/bids/derivatives/atlas/Schaefer2018_400Parcels_17Networks_order_FSLMNI152_2mm.nii.gz \
-    -l /home/james/bids/derivatives/atlas/schaefer_parcel-400_network-17.tsv \
-    -w /home/james/bids/derivatives/betaSeries/work_n17 \
-    -c WhiteMatter CSF Cosine01 Cosine02 Cosine03 Cosine04 Cosine05 Cosine06 Cosine07 \
-    --nthreads 32 \
-    --variant AROMAnonaggr \
-    --hrf-model 'glover + derivative + dispersion'
