### 2019-07-03 97d429e
make code of conduct an rst file
