+++ docs/index.rst
@@ -6,11 +6,12 @@ Contents
 .. toctree::
    :maxdepth: 2
 
-   betaseries
    installation
    usage
    auto_examples/index
    workflows
+   betaseries
+   code_of_conduct
    reference/index
    contributing
    authors
