### 2019-07-03 6c619ad
remove automodule documentation
