### 2019-07-03 710b860
fix syntax warnings
