### 2019-07-03 de01601
add more info on -w help string #133
