### 2019-07-03 HBClab/NiBetaSeries#126 @jdkent
For the example you gave, (`Yes/No`) to (1/0), I do not believe that was strictly necessary, and I can remove that step from the notebook. In general however, I am more inclined to try to have a faithful representation of the [bids specification](https://bids-specification.readthedocs.io/en/stable/04-modality-specific-files/05-task-events.html) and not provide flexibility beyond what the specification allows.
