+++ docs/FAQ.rst
@@ -1,0 +1,14 @@
+.. _faq:
+
+==========================
+Frequently Asked Questions
+==========================
+
+
+Does NiBetaSeries work with output from old fmriprep (< v1.2.0)?
+----------------------------------------------------------------
+Yes, NiBetaSeries will currently work with older version of fmriprep.
+However, this will not continue to be the case.
+To rename your files, you can use renameOldFMRIPREP_
+
+.. _renameOldFMRIPREP: https://github.com/HBClab/renameOldFMRIPREP
