### 2019-07-03 8834e3f
add references/add word parameter estimate #134, #139
