+++ docs/references.bib
@@ -15,24 +15,16 @@ url = {http://journal.frontiersin.org/Article/10.3389/fnsys.2015.00126/abstract}
 volume = {9},
 year = {2015}
 }
-@article{Esteban2018a,
-abstract = {Preprocessing of functional MRI (fMRI) involves numerous steps to clean and standardize data before statistical analysis. Generally, researchers create ad hoc preprocessing workflows for each new dataset, building upon a large inventory of tools available for each step. The complexity of these workflows has snowballed with rapid advances in MR data acquisition and image processing techniques. We introduce fMRIPrep, an analysis-agnostic tool that addresses the challenge of robust and reproducible preprocessing for task-based and resting fMRI data. FMRIPrep automatically adapts a best-in-breed workflow to the idiosyncrasies of virtually any dataset, ensuring high-quality preprocessing with no manual intervention. By introducing visual assessment checkpoints into an iterative integration framework for software-testing, we show that fMRIPrep robustly produces high-quality results on a diverse fMRI data collection comprising participants from 54 different studies in the OpenfMRI repository. We review the distinctive features of fMRIPrep in a qualitative comparison to other preprocessing workflows. We demonstrate that fMRIPrep achieves higher spatial accuracy as it introduces less uncontrolled spatial smoothness than one commonly used preprocessing tool. FMRIPrep has the potential to transform fMRI research by equipping neuroscientists with a high-quality, robust, easy-to-use and transparent preprocessing workflow which can help ensure the validity of inference and the interpretability of their results.},
-archivePrefix = {arXiv},
-arxivId = {1201.0490},
-author = {Esteban, Oscar and Markiewicz, Christopher J and Blair, Ross W and Moodie, Craig A and Ayse, I and Erramuzpe, Asier and Kent, James D and Goncalves, Mathias and DuPre, Elizabeth and Snyder, Madeleine and Oya, Hiroyuki and Ghosh, Satrajit S and Wright, Jessey and Durnez, Joke and Poldrack, Russell A and Gorgolewski, Krzysztof Jacek and Isik, Ayse Ilkay and {Erramuzpe Aliaga}, Asier and Kent, James D and Goncalves, Mathias and DuPre, Elizabeth and Snyder, Madeleine and Oya, Hiroyuki and Ghosh, Satrajit S and Wright, Jessey and Durnez, Joke and Poldrack, Russell A and Gorgolewski, Krzysztof Jacek},
-doi = {10.1101/306951},
-eprint = {1201.0490},
-isbn = {1532-4435},
-issn = {1548-7091},
-journal = {bioRxiv},
-number = {January},
-pages = {1--20},
-pmid = {30532080},
-publisher = {Springer US},
-title = {{FMRIPrep: a robust preprocessing pipeline for functional MRI}},
-url = {http://biorxiv.org/content/early/2018/04/25/306951.abstract},
+@article{Esteban2019,
+author = {Esteban, Oscar and Markiewicz, Christopher J and Blair, Ross W and Moodie, Craig A and Ayse, I and Erramuzpe, Asier and Kent, James D and Goncalves, Mathias and Dupre, Elizabeth and Snyder, Madeleine and Oya, Hiroyuki and Ghosh, Satrajit S and Wright, Jessey and Durnez, Joke and Poldrack, Russell A and Gorgolewski, Krzysztof J},
+doi = {10.1038/s41592-018-0235-4},
+file = {:media/Data/Downloads/fmriprep.pdf:pdf},
+journal = {Nature methods},
+number = {1},
+pages = {111--116},
+title = {{FMRIPrep : a robust preprocessing pipeline for functional MRI}},
 volume = {16},
-year = {2018}
+year = {2019}
 }
 @article{Rissman2004,
[…]
