### 2019-07-03 32a896e
make language clearer and remove reference to high pass #130
