### 2019-07-03 d7677d1
remove step changing yes/no to 1/0 ref #126
