### 2019-07-03 c1b067f
update README.rst to be clear/concise
