+++ MANIFEST.in
@@ -14,11 +14,8 @@ include *.json
 include *.enc
 include requirements.txt
 include requirements-dev.txt
-include AUTHORS.rst
-include CHANGELOG.rst
-include CONTRIBUTING.rst
 include LICENSE
-include README.rst
+include *.rst
 
 include tox.ini .travis.yml
 recursive-include tests *.py
