### 2019-07-03 9c26221
[DOC] Clarify Documentation (#156)
* make code of conduct an rst file

* remove automodule documentation

* reformat readme

* fix syntax warnings

* rename Running NiBetaSeries to Tutorials

* change participant_label to participant-label

* add more info on -w help string #133

* add faq ref: #132

* add references/add word parameter estimate #134, #139

* make it known what options are required #124

* add nature fmriprep reference

* make beta==parameter clearer #134

* make language clearer and remove reference to high pass #130

* remove step changing yes/no to 1/0 ref #126

* update REA
[…]
