+++ examples/README.txt
@@ -1,4 +1,4 @@
-How to run NiBetaSeries
-=======================
+Tutorials
+=========
 
 Below is an example run through of NiBetaSeries
