+++ docs/reference/nibetaseries.rst
@@ -1,9 +1,0 @@
-nibetaseries
-============
-
-.. testsetup::
-
-    from nibetaseries import *
-
-.. automodule:: nibetaseries
-    :members:
