### 2019-07-03 #1295 @jdkent
Update:
I have at least touched on every issue (except adding a binder link), which I am currently deriving a solution for. Thank you for the comments @snastase, I think the documentation is much improved now. @snastase, could you take another look at the [joss issues](https://github.com/HBClab/NiBetaSeries/issues?utf8=%E2%9C%93&q=label%3Ajoss+) to see if I adequately covered your comments?
