### 2019-07-04 2fce650
move dependencies into setup.py
