+++ requirements.txt
@@ -1,11 +1,0 @@
-nipype==1.1.5
-pybids==0.6.4
-nibabel==2.3.0
-nistats==0.0.1b
-nilearn==0.4.2
-pandas==0.24.0
-numpy==1.14.5
-duecredit==0.6.4
-scikit-learn==0.19.2
-matplotlib==2.2.4
-mne==0.18.1
