### 2019-07-04 a1b905f
update tox with new install strategy
