+++ tox.ini
@@ -11,6 +11,8 @@ envlist =
 skip_missing_interpreters = True
 
 [testenv]
+extras =
+    test
 basepython = 
     py35: {env:TOXPYTHON:python3.5}
     py36: {env:TOXPYTHON:python3.6}
@@ -21,17 +23,13 @@ setenv =
     PYTHONUNBUFFERED=yes
 passenv =
     *
-usedevelop = false
-deps =
-    pytest
-    pytest-travis-fold
-    pytest-cov
 commands =
     {posargs:pytest --cov --cov-report=term-missing}
 
 [testenv:docs]
-deps =
-    -r{toxinidir}/docs/requirements.txt
+extras =
+    doc
+    nb
 commands =
     sphinx-build {posargs:-E} -T -D language=en docs docs/_build/html
 
