### 2019-07-04 09fe50b
update manifest
