### 2019-07-04 8789431
remove unused import and remove numpy required version
