### 2019-07-04 6888bd7
[MAINT] Change Install Strategy (#157)
* move dependencies into setup.py

* update tox with new install strategy

* update manifest

* remove unused import and remove numpy required version
