+++ docs/requirements.txt
@@ -1,9 +1,0 @@
-sphinx>=1.3
-sphinx_rtd_theme
-sphinx-argparse
-sphinx-gallery
-sphinxcontrib-bibtex
-matplotlib
-seaborn
-pillow
--e .
