### 2019-07-04 99e0cf0
add binder build step
