### 2019-07-04 a3208fb
add link to binder in the tutorial page
