+++ examples/README.txt
@@ -1,4 +1,0 @@
-Tutorials
-=========
-
-Below is an example run through of NiBetaSeries
