### 2019-07-04 987a73c
[ENH] add binder (#158)
* add binder build step

* add binder badge

* add link to binder in the tutorial page

* add comments and only build on master
