+++ examples/README.rst
@@ -1,0 +1,9 @@
+Tutorials
+=========
+
+Below is an example run through of NiBetaSeries.
+Try "Running NiBetaSeries" with Binder_.
+You will run jupyter through your browser and you will not have
+to install any software.
+
+.. _Binder: https://mybinder.org/v2/gh/HBClab/NiBetaSeries.git/binder?filepath=%2Fbinder%2Fplot_run_nibetaseries.ipynb
