+++ .circleci/config.yml
@@ -35,20 +35,20 @@ jobs:
           no_output_timeout: 60m
           command: |
             docker build \
-              --cache-from=hbclab/nibetaseries \
-              -t hbclab/nibetaseries:latest \
+              --cache-from=hbclab/nibetaseries:unstable \
+              -t hbclab/nibetaseries:unstable \
               --build-arg BUILD_DATE=$(date -u +"%Y-%m-%dT%H:%M:%SZ") \
               --build-arg VCS_REF=$(git rev-parse --short HEAD) .
       - run:
           name: Test Docker image
           command: |
-            docker run hbclab/nibetaseries:latest nibs --help
+            docker run hbclab/nibetaseries:unstable nibs --help
       - run:
           name: Docker save
           no_output_timeout: 30m
           command: |
             mkdir -p /tmp/cache
-            docker save hbclab/nibetaseries:latest | pigz -3 > /tmp/cache/docker.tar.gz
+            docker save hbclab/nibetaseries:unstable | pigz -3 > /tmp/cache/docker.tar.gz
       
       - persist_to_workspace:
           root: /tmp
@@ -58,7 +58,64 @@ jobs:
           key: docker-v1-{{ .Branch }}-{{ .Revision }}
           paths:
             - /tmp/cache/docker.tar.gz
-  
+
+  build_binder:
+    machine:
+      image: circleci/classic:201711-01
+    working_directory: ~/binderBuild
+    steps:
+      - add_ssh_keys:
+          fingerprints:
+            - "b9:b2:c0:2a:28:53:25:bb:b9:c0:4b:ca:13:96:b1:19"
+      - checkout
+      - attach_workspace:
+          at: /tmp
+      - restore_cache:
+          keys:
+            - docker-v1-{{ .Branch }}-{{ .Revision }}
+      - run:
+          name: Load Docker image layer cache
+          no_output_timeout: 30m
+          command: |
+            docker info
+            set +o pipefail
+            if [ -f /tmp/cache/docker.tar.gz ]; then
+              sudo apt update && sudo apt -y install pigz
+              pigz -d --stdout /tmp/cache/docker.tar.gz | docker load
+              docker images
+            fi
+      - run:
+          name: build binder
+          command: |
+            # setup git configuration for pushing to binder branch
+            cd ~/binderBuild
+            git config user.email "binder@binder.com"
+            git config user.name "binder bot"
+            git remote add upstream https://github.com/HBClab/NiBetaSeries.git
+            git checkout -b binder
+            git pull . ${CIRCLE_BRANCH}
+
+            # make directory to place requirements.txt to install nibetaseries
+            # inside binder
+            mkdir -p ~/binderBuild/binder
+            echo ".[nb]" > ~/binderBuild/binder/requirements.txt
+
+            # make the .ipynb file for binder
+            docker run -it \
+              --rm=false \
+              --name binder \
+              -w /src/nibetaseries \
+              hbclab/nibetaseries:unstable \
+              bash -c "source activate neuro_py36 && \
+                       pip install .[doc] && \
+                       sphx_glr_python_to_jupyter.py ./examples/plot_run_nibetaseries.py"
+            # copy the ipynb into the binder folder of the repository
+            docker cp binder:/src/nibetaseries/examples/plot_run_nibetaseries.ipynb ~/binderBuild/binder
+            # push the notebook to the binder branch
+            git add binder/*
+            git commit -m 'add binder files'
+            git push -fq upstream binder
+
   tox_docker:
     machine:
       image: circleci/classic:201711-01
@@ -85,7 +142,7 @@ jobs:
           command: |
             docker run -it \
               -w /src/nibetaseries \
[…]
