### 2019-07-04 088f746
fix tagging/pushing docker images (#160)
