### 2019-07-26 81fc458
Minor changes to documentation text
