+++ docs/betaseries.rst
@@ -4,80 +4,77 @@
 Betaseries
 ==========
 
-This covers the conceptual background of betaseries,
-did you mean to check out usage which talks more about
-how to use NiBetaSeries?
-:ref:`usage`
+This covers the conceptual background of betaseries analysis;
+check out :ref:`usage` for more on how to use NiBetaSeries.
 
 Introduction
 ------------
-Betaseries track the event-to-event modelled hemodynamic fluctuations
-that occur within task functional magnetic resonance imaging (task fMRI).
+Betaseries track the event-to-event hemodynamic fluctuations
+modeled in task functional magnetic resonance imaging (task fMRI).
 Betaseries fills an important analytical gap between measuring hemodynamic
-fluctuations in resting state fMRI and measuring regional activations
+fluctuations in resting-state fMRI and measuring regional activations
 via cognitive subtraction in task fMRI.
 
 Conceptual Background
 ---------------------
 Jesse Rissman :cite:`d-Rissman2004` was the first to publish on betaseries
-correlations describing their usage in a working memory task.
+correlations, describing their usage in a working memory task.
 In this task, participants saw a cue, a delay, and a probe, all occurring
 within a short time period.
-A cue was presented for one second, a delay occurred for seven seconds,
+The cue was presented for one second, a delay occurred for seven seconds,
 and a probe was presented for one second.
-Given the Blood Oxygen-Level Dependent (BOLD) response
+Given that the blood-oxygen-level-dependent (BOLD) response
 takes approximately six seconds to reach its peak, and generally takes over
 20 seconds to return to baseline, we can begin to see a problem.
 The events within the events occur too close to each other to discern what
 brain responses are related to encoding the cue, the delay, or the probe.
 To discern how the activated brain regions form networks, Rissman
-created betaseries correlations.
+computed betaseries correlations.
 Instead of having a single regressor to describe all the cue events,
 a single regressor for all the delay events, and a single regressor for all the
 probe events (as is done in traditional task analysis),
 there is an individual regressor for every event in the experiment.
 For example, if your experiment has 40 events, each with a cue, delay, and
 probe event, the model will have a total of 120 regressors, fitting a beta
-(i.e. parameter) estimate for each event.
-Once you calculate a beta estimate for each event of a given event type
-(e.g. cue), at the end you will have four dimensional dataset where each volume
-represents the beta estimates for a particular event, and each voxel represents
-a specific beta estimate.
+(i.e., parameter) estimate for each event.
+Once you calculate a beta estimate for each event of a given type
+(e.g., cue), you will have a four-dimensional dataset where each volume
+represents the beta estimates for a particular event.
 
-Having one regressor per event in a single model is known as least squares all.
+Having one regressor per event in a single model is known as "least squares all".
 This method, however, has limitations in the context of fast event related
-designs (e.g. designs where the events occur between 3-6
+designs (e.g., designs where the events occur between 3-6
 seconds apart on average).
 Since each event has its own regressor, events that occur very close in time
 are colinear (e.g. are very overlapping).
 Benjamin Turner :cite:`d-Turner2012a` derived a solution for
 the high colinearity observed in least sqaures all by using another
-type of regression known as least squares separate.
-Instead of having one General Linear Model (GLM) with a regressor per event,
+type of regression known as "least squares separate".
+Instead of having one general linear model (GLM) with a regressor per event,
 least squares separate implements a GLM per event with only two regressors:
 1) one for the event of interest, and 2) one for every other event in the
 experiment.
 This process reduces the colinearity of the regressors and creates a more valid
 estimate of how each regressor fits the data.
-NiBetaSeries uses least squares separate and is thus optimized
+NiBetaSeries uses "least squares separate" and is thus optimized
 for fast event related designs.
 
[…]
