+++ docs/FAQ.rst
@@ -5,10 +5,10 @@ Frequently Asked Questions
 ==========================
 
 
-Does NiBetaSeries work with output from old fmriprep (< v1.2.0)?
+Does NiBetaSeries work with output from old fMRIPrep (< v1.2.0)?
 ----------------------------------------------------------------
-Yes, NiBetaSeries will currently work with older version of fmriprep.
+Yes, NiBetaSeries will currently work with older version of fMRIPrep.
 However, this will not continue to be the case.
-To rename your files, you can use renameOldFMRIPREP_
+To rename your files, you can use renameOldFMRIPREP_.
 
 .. _renameOldFMRIPREP: https://github.com/HBClab/renameOldFMRIPREP
