### 2019-07-26 #1295 @snastase
@jdkent sorry the delay—finally got to work back through this. I think this is greatly improved. I went through the documentation and made a [PR](https://github.com/HBClab/NiBetaSeries/pull/163) with minor edits.

I'm satisfied with this but am providing a few additional comments:

1. You might include an equation in the math documentation for LSS to go along with your python pseudocode; e.g.:
```
Y = X_1\beta_1 + X_n\ne1\beta_n\ne1 + \epsilon,
Y = X_2\beta_2 + X_n\ne2\beta_n\ne2 + \epsilon, ...
```

2. Important not to conflate "resting-state", which is a task paradigm, with the correlation-based functional connectivity analyses typically run on resting-state data. I tried to make this a bit clearer in the documentation.

3. I confirmed that tutorial runs on Binder. For whatever reason, the link to Binder in on the Tutorials documentation points a kind of off-putting 404 error until the Binder session initializes. (Note also that the References are not rendering properly at the bottom of the Jupyter notebook).

4. Do you really want to introduce General**ized** Linear Models in the PDF? Not sure this (and the discussion of e.g. non-normality) is really representative of how GLMs are typically used in fMRI. I would also not say the "peak of the gamma curve is determined by the beta coefficient".

5. I would probably the de-emphasize the importance of the "atlas parcellation" in the PDF. Using an atlas to compute inter-areal correlations is only one particular strategy. For example, you could also compute beta series per voxel and then use ICA to identify "networks" rather than using an a prior parcellation.

6. In general, I would do a final pass through the PDF and make sure it's concise as possible.

I would also take out unrelated bits like "voxel is shortened form of volumetric pixel" to make the PDF as short and direct as possible.

I'm curious why you would compute beta series per voxel then average them within each ROI rather than the much less computationally-intensive approach of averaging the time series in each ROI then computing beta series—but this is not a critical issue.
