### 2019-07-28 a235e09
Minor changes to documentation text (#163)
- @snastase edits to documentation to improve clarity
