+++ README.rst
@@ -4,7 +4,7 @@
 NiBetaSeries
 ============
 
-If you are viewing this file through github, please see our
+If you are viewing this file on GitHub, please see our
 `readthedocs page <https://nibetaseries.readthedocs.io>`_
 for links to render properly.
 
@@ -69,9 +69,9 @@ What is NiBetaSeries?
 ---------------------
 NiBetaSeries is BIDS_-compatible `application <https://bids-apps.neuroimaging.io/>`_
 that calculates betaseries correlations.
-In brief, a beta coefficient (i.e. parameter estimate) is calculated
-for each trial (or event) resulting in a series of betas
-that can be used to correlate regions of interest with each other.
+In brief, a beta coefficient (i.e., parameter estimate) is calculated
+for each trial (or event) resulting in a series of betas ("betaseries")
+that can be correlated across regions of interest.
 
 Why should I use it?
 --------------------
@@ -83,27 +83,27 @@ Betaseries can then be used for correlations/classifications and
 a multitude of other analyses.
 While a couple alternatives exist (pybetaseries_ and BASCO_), NiBetaSeries
 is the only application to interface with BIDS_ organized data with the goal
-of providing a commandline application experience like fmriprep_.
+of providing a command-line application experience like fMRIPrep_.
 
 What does NiBetaSeries give me?
 -------------------------------
-Currently NiBetaSeries returns symmetric r-z transformed correlation
+Currently NiBetaSeries returns symmetric z-transformed correlation
 matrices, with an entry for each parcel defined in the atlas.
 Soon, NiBetaSeries will also return the betaseries images themselves,
-so you can take them and flexibly apply another analysis method.
+so you can flexibly apply additional analysis methods.
 
-.. note:: The betas (i.e. parameter estimates) are generated using Least Squares Separate.
+.. note:: The betas (i.e., parameter estimates) are generated using the "Least Squares Separate" procedure.
     Please read the :ref:`betaseries` page for more background information.
     There are plans to support Least Squares All in future iterations.
 
 What do I need to run NiBetaSeries?
 -----------------------------------
 NiBetaSeries takes BIDS_ and preprocessed data as input that satisfy the
-`BIDS deriviatives specification <http://bit.ly/2vKeKcp>`_.
-In practical terms, NiBetaSeries uses the output of fmriprep_,
+`BIDS derivatives specification <http://bit.ly/2vKeKcp>`_.
+In practical terms, NiBetaSeries uses the output of fMRIPrep_,
 a great BIDS-compatible preprocessing tool.
 NiBetaSeries requires the input and the atlas to already
-be in the same space (e.g. MNI space).
+be in the same space (e.g., MNI space).
 For more details, see :ref:`usage` and the tutorial
 (:ref:`sphx_glr_auto_examples_plot_run_nibetaseries.py`)
 
@@ -124,4 +124,4 @@ Please check out their pages and support the developers.
 .. _BASCO: https://www.nitrc.org/projects/basco/
 .. _pybetaseries: https://github.com/poldrack/pybetaseries
 .. _BIDS: http://bids.neuroimaging.io/
-.. _fmriprep: http://fmriprep.readthedocs.io/en/latest/
+.. _fMRIPrep: http://fmriprep.readthedocs.io/en/latest/
