+++ examples/README.rst
@@ -3,7 +3,8 @@ Tutorials
 
 Below is an example run through of NiBetaSeries.
 Try "Running NiBetaSeries" with Binder_.
-You will run jupyter through your browser and you will not have
-to install any software.
+It will run Jupyter through your browser and you will not have
+to install any software (it may take a few seconds for the Binder session to
+launch).
 
 .. _Binder: https://mybinder.org/v2/gh/HBClab/NiBetaSeries.git/binder?filepath=%2Fbinder%2Fplot_run_nibetaseries.ipynb
