### 2019-07-28 HBClab/NiBetaSeries#164 @jdkent
Add Math equation to LSS
[original](https://github.com/openjournals/joss-reviews/issues/1295#issuecomment-515579870)

> You might include an equation in the math documentation for LSS to go along with your python pseudocode; e.g.:
Y = X_1\beta_1 + X_n\ne1\beta_n\ne1 + \epsilon,
Y = X_2\beta_2 + X_n\ne2\beta_n\ne2 + \epsilon, ...
