### 2019-07-28 HBClab/NiBetaSeries#165 @jdkent
conflation of resting-state paradigm and correlation fc method
[original](https://github.com/openjournals/joss-reviews/issues/1295#issuecomment-515579870)

> Important not to conflate "resting-state", which is a task paradigm, with the correlation-based functional connectivity analyses typically run on resting-state data. I tried to make this a bit clearer in the documentation.

Read through documentation to make sure is clear
ref #163
