### 2019-07-28 HBClab/NiBetaSeries#166 @jdkent
jupyter binder: misleading 404
[original](https://github.com/openjournals/joss-reviews/issues/1295#issuecomment-515579870)
> I confirmed that tutorial runs on Binder. For whatever reason, the link to Binder in on the Tutorials documentation points a kind of off-putting 404 error until the Binder session initializes. (Note also that the References are not rendering properly at the bottom of the Jupyter notebook).
