### 2019-07-28 HBClab/NiBetaSeries#167 @jdkent
misrepresentation of Generalized Linear Models
[original](https://github.com/openjournals/joss-reviews/issues/1295#issuecomment-515579870)
> Do you really want to introduce Generalized Linear Models in the PDF? Not sure this (and the discussion of e.g. non-normality) is really representative of how GLMs are typically used in fMRI. I would also not say the "peak of the gamma curve is determined by the beta coefficient".
