### 2019-07-28 HBClab/NiBetaSeries#168 @jdkent
be analysis agnostic in the joss paper
[original](https://github.com/openjournals/joss-reviews/issues/1295#issuecomment-515579870)
> I would probably the de-emphasize the importance of the "atlas parcellation" in the PDF. Using an atlas to compute inter-areal correlations is only one particular strategy. For example, you could also compute beta series per voxel and then use ICA to identify "networks" rather than using an a prior parcellation.
