### 2019-07-28 HBClab/NiBetaSeries#169 @jdkent
conciseness of joss paper
[original](https://github.com/openjournals/joss-reviews/issues/1295#issuecomment-515579870)
> In general, I would do a final pass through the PDF and make sure it's concise as possible. For example, I would take out unrelated bits like "voxel is shortened form of volumetric pixel" to make the PDF as short and direct as possible.
