### 2019-07-28 HBClab/NiBetaSeries#170 @jdkent
inefficient computations
[original](https://github.com/openjournals/joss-reviews/issues/1295#issuecomment-515579870)
> I'm curious why you would compute beta series per voxel then average them within each ROI rather than the much less computationally-intensive approach of averaging the time series in each ROI then computing beta series—but this is not a critical issue.
