### 2019-08-13 f84b4eb
add pypiwin32 as dependency if os is windows
