### 2019-08-13 5a5031a
[FIX] add readthedocs config file (#174)
* add readthedocs config

* add yml files to manifest
