+++ .readthedocs.yml
@@ -1,0 +1,9 @@
+version: 2
+
+python:
+  install:
+    - method: pip
+      path: .
+      extra_requirements:
+        - doc
+        - nb
