### 2019-08-13 d17244f
remove .git from the url
