+++ examples/README.rst
@@ -7,4 +7,4 @@ It will run Jupyter through your browser and you will not have
 to install any software (it may take a few seconds for the Binder session to
 launch).
 
-.. _Binder: https://mybinder.org/v2/gh/HBClab/NiBetaSeries.git/binder?filepath=%2Fbinder%2Fplot_run_nibetaseries.ipynb
+.. _Binder: https://mybinder.org/v2/gh/HBClab/NiBetaSeries/binder?filepath=%2Fbinder%2Fplot_run_nibetaseries.ipynb
