### 2019-08-13 4a1bdee
remove .git from the url (#175)
