### 2019-08-19 #1295 @arokem
@jdkent : how are things going? When do you anticipate finishing your current revisions?
