### 2019-08-19 #1295 @jdkent
I hope to have the current revisions done by Friday this week.
