### 2019-08-28 f59b7e4
treat warnings as errors in sphinx build
