+++ tox.ini
@@ -31,7 +31,7 @@ extras =
     doc
     nb
 commands =
-    sphinx-build {posargs:-E} -T -D language=en docs docs/_build/html
+    sphinx-build {posargs:-E} -W -T -D language=en docs docs/_build/html
 
 [testenv:check]
 deps =
