### 2019-08-28 2b0b577
bolster math and code documentation
