### 2019-08-28 7a4ef2b
fix explanation of glm
