+++ src/nibetaseries/interfaces/nistats.py
@@ -80,7 +80,8 @@ class BetaSeries(NistatsBaseInterface, SimpleInterface):
         # initialize dictionary to contain trial estimates (betas)
         beta_maps = {}
         design_matrix_collector = {}
-        for target_trial_df, trial_type, trial_idx in _lss_events_iterator(self.inputs.events_file):
+        for target_trial_df, trial_type, trial_idx in \
+                _lss_events_iterator(self.inputs.events_file):
 
             # fit the model for the target trial
             model.fit(self.inputs.bold_file,
@@ -98,15 +99,17 @@ class BetaSeries(NistatsBaseInterface, SimpleInterface):
                 beta_maps[trial_type] = [beta_map]
 
         # make a beta series from each beta map list
-        beta_series_template = os.path.join(runtime.cwd, 'betaseries_trialtype-{trial_type}.nii.gz')
+        beta_series_template = os.path.join(runtime.cwd,
+                                            'betaseries_trialtype-{trial_type}.nii.gz')
         # collector for the betaseries files
         beta_series_lst = []
         for t_type, betas in beta_maps.items():
             size_check = len(betas)
             if size_check < 3:
-                logging.warning('At least 3 trials are needed '
-                                'for a beta series: {trial_type} has {num}'.format(trial_type=t_type,
-                                                                                   num=size_check))
+                logging.warning(
+                    'At least 3 trials are needed '
+                    'for a beta series: {trial_type} has {num}'.format(trial_type=t_type,
+                                                                       num=size_check))
             else:
                 beta_series = nib.funcs.concat_images(betas)
                 nib.save(beta_series, beta_series_template.format(trial_type=t_type))
