+++ .joss/paper.bib
@@ -111,4 +111,29 @@ url = {https://doi.org/10.1371/journal.pcbi.1005209},
 volume = {13},
 year = {2017}
 }
+@article{Cole2010,
+abstract = {The last 15 years have witnessed a steady increase in the number of resting-state functional neuroimaging studies. The connectivity patterns of multiple functional, distributed, large-scale networks of brain dynamics have been recognised for their potential as useful tools in the domain of systems and other neurosciences. The application of functional connectivity methods to areas such as cognitive psychology, clinical diagnosis and treatment progression has yielded promising preliminary results, but is yet to be fully realised. This is due, in part, to an array of methodological and interpretative issues that remain to be resolved. We here present a review of the methods most commonly applied in this rapidly advancing field, such as seed-based correlation analysis and independent component analysis, along with examples of their use at the individual subject and group analysis levels and a discussion of practical and theoretical issues arising from this data ‘explosion'. We describe the similarities and differences across these varied statistical approaches to processing resting-state FMRI signals, and conclude that further technical optimisation and experimental refinement is required in order to fully delineate and characterise the gross complexity of the human neural functional architecture.},
+author = {Cole, David and Smith, Stephen and Beckmann, Christian},
+doi = {10.3389/fnsys.2010.00008},
+issn = {1662-5137},
+journal = {Frontiers in Systems Neuroscience},
+pages = {8},
+title = {{Advances and pitfalls in the analysis and interpretation of resting-state FMRI data}},
+url = {https://www.frontiersin.org/article/10.3389/fnsys.2010.00008},
+volume = {4},
+year = {2010}
+}
+@article{vanWijk2010,
+abstract = {Graph theory is a valuable framework to study the organization of functional and anatomical connections in the brain. Its use for comparing network topologies, however, is not without difficulties. Graph measures may be influenced by the number of nodes (N) and the average degree (k) of the network. The explicit form of that influence depends on the type of network topology, which is usually unknown for experimental data. Direct comparisons of graph measures between empirical networks with different N and/or k can therefore yield spurious results. We list benefits and pitfalls of various approaches that intend to overcome these difficulties. We discuss the initial graph definition of unweighted graphs via fixed thresholds, average degrees or edge densities, and the use of weighted graphs. For instance, choosing a threshold to fix N and k does eliminate size and density effects but may lead to modifications of the network by enforcing (ignoring) non-significant (significant) connections. Opposed to fixing N and k, graph measures are often normalized via random surrogates but, in fact, this may even increase the sensitivity to differences in N and k for the commonly used clustering coefficient and small-world index. To avoid such a bias we tried to estimate the N,k-dependence for empirical networks, which can serve to correct for size effects, if successful. We also add a number of methods used in social sciences that build on statistics of local network structures including exponential random graph models and motif counting. We show that none of the here-investigated methods allows for a reliable and fully unbiased comparison, but some perform better than others.},
+author = {van Wijk, Bernadette C M and Stam, Cornelis J and Daffertshofer, Andreas},
+doi = {10.1371/journal.pone.0013701},
+journal = {PLOS ONE},
+number = {10},
+pages = {1--13},
+publisher = {Public Library of Science},
+title = {{Comparing Brain Networks of Different Size and Connectivity Density Using Graph Theory}},
+url = {https://doi.org/10.1371/journal.pone.0013701},
+volume = {5},
+year = {2010}
+}
 
