### 2019-08-28 3d54765
[ENH] add more example code to documentation (#177)
* move max line number to 99

* treat warnings as errors in sphinx build

* fix line indentations

* bolster math and code documentation

* fix line length errors
