+++ src/nibetaseries/workflows/base.py
@@ -17,10 +17,13 @@ from nipype.interfaces import utility as niu
 from bids import BIDSLayout
 
 
-def init_nibetaseries_participant_wf(atlas_img, atlas_lut, bids_dir,
-                                     derivatives_pipeline_dir, exclude_variant_label, hrf_model, low_pass,
-                                     output_dir, run_label, selected_confounds, session_label, smoothing_kernel,
-                                     space_label, subject_list, task_label, variant_label, work_dir):
+def init_nibetaseries_participant_wf(
+    atlas_img, atlas_lut, bids_dir,
+    derivatives_pipeline_dir, exclude_variant_label, hrf_model, low_pass,
+    output_dir, run_label, selected_confounds, session_label, smoothing_kernel,
+    space_label, subject_list, task_label, variant_label, work_dir,
+        ):
+
     """
     This workflow organizes the execution of NiBetaSeries, with a sub-workflow for
     each subject.
@@ -118,11 +121,14 @@ def init_nibetaseries_participant_wf(atlas_img, atlas_lut, bids_dir,
     return nibetaseries_participant_wf
 
 
-def init_single_subject_wf(atlas_img, atlas_lut, bold_metadata_list, brainmask_list, confound_tsv_list,
-                           events_tsv_list, hrf_model, low_pass, name, output_dir, preproc_img_list,
-                           selected_confounds, smoothing_kernel):
+def init_single_subject_wf(
+    atlas_img, atlas_lut, bold_metadata_list, brainmask_list, confound_tsv_list,
+    events_tsv_list, hrf_model, low_pass, name, output_dir, preproc_img_list,
+    selected_confounds, smoothing_kernel
+        ):
     """
-    This workflow completes the generation of the betaseries files and the calculation of the correlation matrices.
+    This workflow completes the generation of the betaseries files
+    and the calculation of the correlation matrices.
     .. workflow::
         :graph2use: orig
         :simple_form: yes
@@ -161,7 +167,8 @@ def init_single_subject_wf(atlas_img, atlas_lut, bold_metadata_list, brainmask_l
         hrf_model : str
             hemodynamic response function used to model the data
         low_pass : float or None
-            low pass filter to apply to bold (in Hertz). Reminder - frequencies _lower_ than this number are kept.
+            low pass filter to apply to bold (in Hertz).
+            Reminder - frequencies _lower_ than this number are kept.
         name : str
             name of the workflow (e.g. ``subject-01_wf``)
         output_dir : str
@@ -227,7 +234,8 @@ def init_single_subject_wf(atlas_img, atlas_lut, bold_metadata_list, brainmask_l
 
     # initialize the betaseries workflow
     betaseries_wf = init_betaseries_wf(hrf_model=hrf_model, low_pass=low_pass,
-                                       selected_confounds=selected_confounds, smoothing_kernel=smoothing_kernel)
+                                       selected_confounds=selected_confounds,
+                                       smoothing_kernel=smoothing_kernel)
 
     # initialize the analysis workflow
     correlation_wf = init_correlation_wf()
@@ -259,7 +267,8 @@ def init_single_subject_wf(atlas_img, atlas_lut, bold_metadata_list, brainmask_l
         (correlation_wf, output_node, [('output_node.correlation_matrix', 'correlation_matrix'),
                                        ('output_node.correlation_fig', 'correlation_fig')]),
         (input_node, ds_correlation_matrix, [('preproc_img', 'source_file')]),
-        (betaseries_wf, ds_correlation_matrix, [('output_node.betaseries_files', 'betaseries_file')]),
+        (betaseries_wf, ds_correlation_matrix, [('output_node.betaseries_files',
[…]
