+++ setup.cfg
@@ -2,12 +2,12 @@
 universal = 1
 
 [flake8]
-max-line-length = 120
+max-line-length = 99
 exclude = */migrations/*
 
 [isort]
 force_single_line = True
-line_length = 120
+line_length = 99
 known_first_party = nibetaseries
 default_section = THIRDPARTY
 forced_separate = test_nibetaseries
