### 2019-08-28 f45e939
[FIX] generalized -> general linear model description (#178)
* fix explanation of glm

* fix grammatical mistake
