### 2019-08-29 #1295 @snastase
> Great! @snastase : could you please confirm that everything has been addressed? I see that all the check-boxes are checked, which means that this is probably ready to go.

@arokem Good to go on my end! @jdkent has addressed my comments.
