### 2019-08-29 #1295 @arokem
Great! Thanks for all your work on this review, @snastase. I concur on my end.

At this point, @jdkent : could you please create an archive of your repository (e.g., using Zenodo) and then post the DOI here. When creating the archive, please make sure that:

- The authors of the archive are the same as the JOSS paper authors
