### 2019-08-29 f8daeb4
add peer herholz to joss paper
