### 2019-08-30 a639071
[REL] v0.3.0 release (#181)
* update changelog

* modify the zenodo files for this release

* add peer herholz to joss paper

* fix peer affiliation

* address peer comments

* clarify sentence by adding brain
