### 2019-08-31 #1295 @jdkent
@arokem here is the doi: 10.5281/zenodo.3382330

I had some trouble with zenodo, but I think I have it all worked out, [the page is here](https://zenodo.org/record/3382330#.XWm5Yx-YU8o)

I second that big thanks to @snastase! and thank you @arokem for keeping us moving along!
