+++ .circleci/config.yml
@@ -216,12 +216,21 @@ workflows:
   build_test_deploy:
     jobs:
       - build_docker
+          filters:
+            tags:
+              only: /^v.*/
       - tox_docker:
           requires:
             - build_docker
+          filters:
+            tags:
+              only: /^v.*/            
       - build_docs:
           requires:
             - build_docker
+          filters:
+            tags:
+              only: /^v.*/
       - build_binder:
           requires:
             - build_docker
@@ -229,6 +238,8 @@ workflows:
             branches:
               only:
                 - master
+            tags:
+              only: /^v.*/
       - deploy_docker:
           context: hbclab
           requires:
@@ -238,4 +249,4 @@ workflows:
             branches:
               only: master
             tags:
-              only: /.*/
+              only: /^v.*/
