### 2019-08-31 62744ba
pin pytest version https://github.com/microsoft/vscode-python/issues/6990
