### 2019-08-31 974b078
update manifest
