### 2019-09-01 #1295 @arokem
Hi @jdkent : I see that the Zenodo submission has two authors (you and Peer), but you are the sole author of the paper. That needs to be consistent. I believe that you should be able to edit the Zenodo submission metadata.
