### 2019-09-01 #1295 @jdkent
I [remade the paper](https://github.com/openjournals/joss-papers/blob/joss.01295/joss.01295/10.21105.joss.01295.pdf) with Peer as a co-author, or were you referencing another document?
