### 2019-09-01 #1295 @arokem
OK. I see that he approved that in https://github.com/HBClab/NiBetaSeries/pull/181.
