### 2019-09-01 #1295 @arokem
@openjournals/joss-eics : I believe that this is now ready for your review
