### 2019-09-01 #1295 @labarba
I see that the latest version tagged on the GitHub repo is v0.3.0, while here we have v0.2.3. Please confirm that we should change the version here to match the tag in the repo. Also, it is marked as a "Pre-release." Are you sure you want that? It generally means the release is unstable—if that were the case, you would not be publishing a paper about the software at this time.

Since I'm sending this back to you with release questions, might as well add minor edits: please capitalize Python and Matlab, use commas after "e.g." and "i.e." and make one final pass at conciseness. For example, you can delete "With the concept of beta under our belt" (excess clutter). The phrase "Neurohackademy is to thank" is grammatically awkward. I'm also a bit unsatisfied with your headings: Introduction, Background, Overview. It seems all headings in the paper could be shuffled around without harm to the narrative, which makes me think they are superfluous. Perhaps just one Background heading and one Software Overview? (We sometimes also see Statement of Need, given that this is one of the [requested topics in our docs](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain). Give it a thought.
