### 2019-09-01 6caa950
address @labarba review comments
