+++ .joss/paper.md
@@ -1,7 +1,7 @@
 ---
 title: 'NiBetaSeries: task related correlations in fMRI'
 tags:
-  - python
+  - Python
   - neuroscience
   - neuroimaging
   - BIDS
@@ -22,97 +22,74 @@ date: 30 January 2019
 bibliography: paper.bib
 ---
 
-# Introduction
+# Background
 
 You may not be aware of it, but your brain is orchestrating a complex ballet of activity while reading this sentence.
-Whether it is following a dot, reading a sentence, interpreting what response is required when presented with a red square; the brain is evaluating input and sending motoric output to perform optimally/efficiently.
-We can measure and analyze this flurry of activity using functional Magnetic Resonance Imaging (fMRI).
+Whether it is following a dot or reading a sentence; the brain is evaluating input and sending motoric output to perform optimally/efficiently.
+We can measure this flurry of activity using functional Magnetic Resonance Imaging (fMRI).
 Traditional fMRI analysis emphasizes what regions are "activated/deactivated" during a task, but does not provide information on which regions are acting in synchrony or are being segregated.
 Knowing the synchronous/segregated brain regions during a task gives insights on the potential organization of the brain.
-NiBetaSeries seeks to fill this gap by providing a method to measure how activation/deactivation
-patterns between brain regions are correlated across trials.
-That is to say if region A and region B both show high activity on trial 1, low activity on trial 2, medium activity on trial 3, and so on, then these two regions are highly correlated.
-
-# Background
+**NiBetaSeries seeks to inform about the organization of the brain by correlating activation/deactivation patterns between brain regions during a task.**
 
 To understand NiBetaSeries we need to answer two questions: what is a "beta" (or parameter estimate) and how can we analyze a series of betas?
 We have already mentioned betas by another name, activation/deactivation.
 The term beta comes from its use in the General Linear Model (GLM), an extension of linear regression.
-fMRI signal evoked by a stimulus follows a relatively stereotyped shape from the Blood Oxygen Level Dependent (BOLD) response, which is best modeled with a double-gamma function.
-The overall amplitude (i.e. activation/deactivation) of the double gamma function is determined by the beta coefficient.
-Thus larger betas mean greater activation, and smaller or negative betas mean less activation or deactivation relative to a baseline.
-Traditional fMRI analysis will group together all the relevant trial types and give them all one beta estimate, where variance between trials is treated as noise.
+fMRI signal evoked by a stimulus follows a relatively stereotyped shape from the Blood Oxygen Level Dependent (BOLD) response, which is reasonably modeled with a double-gamma function.
+The overall amplitude (i.e., activation/deactivation) of the double gamma function is determined by the beta coefficient.
+Thus larger betas mean greater activation and smaller or negative betas mean less activation or deactivation relative to a baseline.
+Traditional fMRI analysis will group together all trials of a single type and give them all one beta estimate, where variance between trials is treated as noise.
 NiBetaSeries, on the other hand, gives each trial its own beta estimate treating the variance between trials as the signal of interest.
-With the concept of a beta under our belt, The final complication is the multiple methods we can use to derive individual beta estimates.
-Two common methods are least squares all (LSA) and least squares separate (LSS) [@Mumford2012;@Turner2012a].
+
+Two common methods for deriving single trial beta estimates are least squares all (LSA) and least squares separate (LSS) [@Mumford2012;@Turner2012a].
 LSA places all the trials in the same GLM, where each trial is a separate predictor.
-LSA works well when the trials are far apart in time since the BOLD response takes a long time to return to baseline.
[…]
