### 2019-09-01 HBClab/NiBetaSeries#184 @jdkent
Address Joss Editor Review Comments
related to #169, I need to update the paper to be more concise and include a statement of need.

See [original comment](https://github.com/openjournals/joss-reviews/issues/1295#issuecomment-526940158)
