### 2019-09-01 #1295 @jdkent
Thanks for noting the pre-release check, that was a mistake.

Thank you for the feedback @labarba, I believe I've addressed your comments in the referenced pull request.

Note, I've merged in [one pull request](https://github.com/HBClab/NiBetaSeries/pull/183) since the making the zenodo doi, with [another on right on it's heels](https://github.com/HBClab/NiBetaSeries/pull/182), so I could just do one more minor release and say `v0.3.1` will be the JOSS release when merging HBClab/NiBetaSeries#184.

The alternative as I see it is to modify the `v0.3.0` release, either way, I would like to change the version from `v0.2.3`.

Let me know if you have any preferences for modifying `v0.3.0` or cutting a new minor release (`v0.3.1`) for the JOSS submission.

Thanks!
