### 2019-09-01 #1295 @labarba
Feel free to increase the patch release, and report the final version number here, so we can add it.
