### 2019-09-02 e80efe9
Retain conditions' trial types.
