### 2019-09-03 b01cd7d
add code-server instructions eliminate broken links misc spelling
