+++ CONTRIBUTING.rst
@@ -4,6 +4,10 @@
 Contributing to NiBetaSeries
 ============================
 
+If you are viewing this file on GitHub, please see our
+`readthedocs page <https://nibetaseries.readthedocs.io>`_
+for links/images to render properly.
+
 **Welcome to the NiBetaSeries repository!**
 
 *We're so excited you're here and want to
@@ -24,15 +28,14 @@ Been here before? Already know what you're looking for in this guide?
 Jump to the following sections:
 
 -  `Joining the BIDS community <#joining-the-bids-community>`__
--  `Check out the BIDS Starter Kit <#check-bids-starter-kit>`__
+-  `Check out the BIDS Starter Kit <#check-out-the-bids-starter-kit>`__
 -  `Get in touch <#get-in-touch>`__
 -  `Contributing through GitHub <#contributing-through-github>`__
--  `Where to start: wiki, code and
-   templates <#where-to-start-wiki-code-and-templates>`__
 -  `Where to start: issue labels <#where-to-start-issue-labels>`__
 -  `Make a change with a pull
    request <#making-a-change-with-a-pull-request>`__
--  `Example pull request <#example-pull-request>`__
+-  `Code Server Development Environment <#code-server-development>`__
+-  `Docker Toolbox Instructions <#docker-toolbox-instructions>`__
 -  `Recognizing contributions <#recognizing-contributions>`__
 
 Joining the BIDS community
@@ -120,7 +123,7 @@ a documentation hosting service.
 Putting it all together, ReadTheDocs is an online documentation hosting service
 that uses sphinx, and sphinx is a documentation generation service that uses
 ReStructuredText to format the content.
-What a mouthfull!
+What a mouthful!
 
 Where to start: issue labels
 ----------------------------
@@ -252,6 +255,10 @@ Example::
     # create the branch on which you will make your issues
     git checkout -b your_issue_branch
 
+If you are making significant changes to the code, you may wish
+to setup a local development environment.
+Checkout the `code-server <#code-server-development>`__ section for details.
+
 5. Run the tests
 ~~~~~~~~~~~~~~~~
 
@@ -259,7 +266,7 @@ When you're done making changes, run all the checks, doc builder and spell check
 First you will install all the development requirements
 for the project with pip::
 
-    pip install requirements-dev.txt
+    pip install .[test]
 
 Then you can run tox by simply typing::
 
@@ -344,6 +351,103 @@ identifier or a one-word description, such as::
 
     myworkflow_lh_wf = init_workflow_wf(name='myworkflow_lh_wf')
     myworkflow_rh_wf = init_workflow_wf(name='myworkflow_rh_wf')
+
+Code Server Development
+-----------------------
+
+To get the best of working with containers and having an interactive
+development environment, we have an experimental setup with `code-server
+<https://github.com/cdr/code-server>`__.
+
+.. Note::
+    We have `a video walking through the process for fmriprep
+    <https://youtu.be/bkZ-NyUaTvg>`__ if you want a visual guide.
+    Hopefully the process for nibetaseries is similar enough for the video
+    to be useful.
+
+1. Build the Docker image
+~~~~~~~~~~~~~~~~~~~~~~~~~
+We assume you have NiBetaSeries cloned in $HOME/projects.
+We will use the ``Dockerfile_devel`` file to build
+our development docker image::
+
+    $ cd $HOME/projects/NiBetaSeries
+    $ docker build -t nibetaseries_devel -f Dockerfile_devel .
+
+2. Run the Docker image
+~~~~~~~~~~~~~~~~~~~~~~~
+We can start a docker container using the image we built (``nibetaseries_devel``)::
+
[…]
