+++ docs/conf.py
@@ -53,6 +53,8 @@ html_sidebars = {
 }
 html_short_title = '%s-%s' % (project, version)
 
+html_static_path = ['_static']
+
 napoleon_use_ivar = True
 napoleon_use_rtype = False
 napoleon_use_param = False
