### 2019-09-03 7aa4646
add documentation for adding git credentials to docker container [build devel]
