### 2019-09-03 5d256f9
make pip install editable
