### 2019-09-03 a3ed83a
add restructuredtext to list of extensions installed
