### 2019-09-03 6adda90
[ENH] add code server (#182)
* add the vscode folder for development settings

* pin pytest version
https://github.com/microsoft/vscode-python/issues/6990

* make docker development file

* update manifest

* add docker toolbox visual instructions

* add code-server instructions
eliminate broken links
misc spelling

* add static file path in conf.py

* add restructeredtext previewer

* change default settings including not showing gitlens intro screen

* add optional devel build [build]

* add documentation for adding git credentials to docker container [build devel]

* make pip install editab
[…]
