+++ Dockerfile
@@ -1,5 +1,5 @@
 # Generated by Neurodocker version 0.5.0
-# Timestamp: 2019-07-02 21:26:54 UTC
+# Timestamp: 2019-08-31 17:22:00 UTC
 # 
 # Thank you for using Neurodocker. If you discover any issues
 # or ways to improve this software, please submit an issue or
@@ -89,7 +89,7 @@ USER neuro
 
 WORKDIR /src/nibetaseries
 
-RUN pip install --no-cache-dir .
+RUN . activate neuro_py36 && pip install --no-cache-dir .[test,dev,doc,nb]
 
 RUN mkdir -p ~/.jupyter && echo c.NotebookApp.ip = \"0.0.0.0\" > ~/.jupyter/jupyter_notebook_config.py
 
@@ -158,7 +158,7 @@ RUN echo '{ \
     \n    ], \
     \n    [ \
     \n      "run", \
-    \n      "pip install --no-cache-dir ." \
+    \n      ". activate neuro_py36 && pip install --no-cache-dir .[test,dev,doc,nb]" \
     \n    ], \
     \n    [ \
     \n      "run", \
