### 2019-09-03 1761e92
move install to cfg/toml
