+++ setup.py
@@ -1,101 +1,25 @@
 #!/usr/bin/env python
 # -*- encoding: utf-8 -*-
-from __future__ import absolute_import
-from __future__ import print_function
-
-import io
-import re
-from glob import glob
-from os.path import dirname, join, basename, splitext
-from setuptools import find_packages
-from setuptools import setup
+"""nibetaseries setup script"""
+import sys
+from setuptools import setup, find_packages
 import versioneer
 
+# Give setuptools a hint to complain if it's too old a version
+# 30.3.0 allows us to put most metadata in setup.cfg
+# Should match pyproject.toml
+# Not going to help us much without numpy or new pip, but gives us a shot
+SETUP_REQUIRES = ['setuptools >= 40.8']
+# This enables setuptools to install wheel on-the-fly
+SETUP_REQUIRES += ['wheel'] if 'bdist_wheel' in sys.argv else []
 
-def read(*names, **kwargs):
-    return io.open(
-        join(dirname(__file__), *names),
-        encoding=kwargs.get('encoding', 'utf8')
-    ).read()
 
+if __name__ == '__main__':
 
-setup(
-    name='nibetaseries',
-    license='MIT license',
-    version=versioneer.get_version(),
-    cmdclass=versioneer.get_cmdclass(),
-    description='BetaSeries Correlations implemented in Nipype',
-    long_description='%s\n%s' % (
-        re.compile('^.. start-badges.*^.. end-badges', re.M | re.S).sub('', read('README.rst')),
-        re.sub(':[a-z]+:`~?(.*?)`', r'``\1``', read('CHANGELOG.rst'))
-    ),
-    author='James Kent',
-    author_email='james-kent@uiowa.edu',
-    url='https://github.com/HBClab/NiBetaSeries',
-    packages=find_packages('src'),
-    package_dir={'': 'src'},
-    py_modules=[splitext(basename(path))[0] for path in glob('src/*.py')],
-    include_package_data=True,
-    zip_safe=False,
-    classifiers=[
-        # complete classifier list: http://pypi.python.org/pypi?%3Aaction=list_classifiers
-        'Development Status :: 1 - Planning',
-        'Environment :: Console',
-        'Intended Audience :: Science/Research',
-        'License :: OSI Approved :: MIT License',
-        'Operating System :: Unix',
-        'Operating System :: POSIX',
-        'Programming Language :: Python',
-        'Programming Language :: Python :: 3',
-        'Programming Language :: Python :: 3.6',
-        'Programming Language :: Python :: Implementation :: CPython',
-        'Programming Language :: Python :: Implementation :: PyPy',
-        'Topic :: Utilities',
-    ],
-    keywords=[
-        'bids',
-        'fmri',
-        'neuroimaging',
-    ],
-    install_requires=[
-        'nipype~=1.1.5',
-        'pybids~=0.6.4',
-        'nibabel~=2.3.0',
-        'nistats~=0.0.1b',
-        'nilearn~=0.4.2',
-        'pandas~=0.24.0',
-        'numpy',
-        'duecredit~=0.6.4',
-        'scikit-learn~=0.19.2',
-        'matplotlib~=2.2.4',
-        'mne~=0.18.1',
-        'pypiwin32; platform_system=="Windows"',
-    ],
-    extras_require={
-        'test': ['tox',
-                 'pytest~=5.0.1',
-                 'pytest-travis-fold',
-                 'pytest-cov'],
-        'dev': ['check-manifest',
-                'flake8',
-                'codecov',
-                'coverage'],
-        'doc': ['sphinx>=1.3',
-                'sphinx_rtd_theme',
-                'sphinx-argparse',
-                'sphinx-gallery',
[…]
