+++ setup.cfg
@@ -1,5 +1,86 @@
-[bdist_wheel]
-universal = 1
+[metadata]
+url = https://github.com/HBClab/NiBetaSeries
+author = James Kent
+author_email = james-kent@uiowa.edu
+maintainer = James Kent
+maintainer_email = james-kent@uiowa.edu
+description = BetaSeries Correlations implemented in Nipype
+long_description = file:long_description.rst
+long_description_content_type = text/x-rst; charset=UTF-8
+licence = MIT
+classifiers =
+    Development Status :: 1 - Planning
+    Environment :: Console
+    Intended Audience :: Science/Research
+    License :: OSI Approved :: MIT License
+    Operating System :: Unix
+    Operating System :: POSIX
+    Programming Language :: Python
+    Programming Language :: Python :: 3.5
+    Programming Language :: Python :: 3.6
+    Programming Language :: Python :: 3.7
+    Topic :: Utilities
+keywords = 
+    bids
+    fmri
+    neuroimaging
+
+[options]
+python_requires = >=3.5
+install_requires = 
+    nipype ~= 1.1.5
+    pybids ~= 0.6.4
+    nibabel ~= 2.3.0
+    nistats ~= 0.0.1b
+    nilearn ~= 0.4.2
+    pandas ~= 0.24.0
+    numpy
+    duecredit ~= 0.6.4
+    scikit-learn ~= 0.19.2
+    matplotlib ~= 2.2.4
+    mne ~= 0.18.1
+    pypiwin32; platform_system=="Windows"
+
+packages_dir =
+    = src
+packages = find:
+
+[options.packages.find]
+where=src
+
+[options.extras_require]
+test =
+    tox
+    pytest ~= 5.0.1
+    pytest-travis-fold
+    pytest-cov
+tests =
+    %(test)s
+dev =
+    check-manifest
+    flake8
+    codecov
+    coverage
+doc = 
+    sphinx >= 1.3
+    sphinx_rtd_theme
+    sphinx-argparse
+    sphinx-gallery
+    sphinxcontrib-bibtex
+docs =
+    %(doc)s
+nb =
+    seaborn
+    pillow
+all =
+    %(test)s
+    %(nb)s
+    %(doc)s
+    %(dev)s
+
+[options.entry_points]
+console_scripts =
+    nibs=nibetaseries.cli.run:main
 
 [flake8]
 max-line-length = 99
@@ -18,5 +99,5 @@ skip = migrations
 VCS = git
 style = pep440-post
 versionfile_source = src/nibetaseries/_version.py
-versionfile_build = nibetaseries/_version.py
+versionfile_build = src/nibetaseries/_version.py
 tag_prefix = v
