### 2019-09-03 16ecd68
add cython as dependency (for py3.5?)
