+++ setup.py
@@ -9,7 +9,7 @@ import versioneer
 # 30.3.0 allows us to put most metadata in setup.cfg
 # Should match pyproject.toml
 # Not going to help us much without numpy or new pip, but gives us a shot
-SETUP_REQUIRES = ['setuptools >= 40.8']
+SETUP_REQUIRES = ['setuptools >= 40.8', 'cython']
 # This enables setuptools to install wheel on-the-fly
 SETUP_REQUIRES += ['wheel'] if 'bdist_wheel' in sys.argv else []
 
