+++ pyproject.toml
@@ -1,2 +1,2 @@
 [build-system]
-requires = ["setuptools >= 40.8.0", "wheel"]
+requires = ["setuptools >= 40.8.0", "wheel", "cython"]
