+++ setup.cfg
@@ -99,5 +99,5 @@ skip = migrations
 VCS = git
 style = pep440-post
 versionfile_source = src/nibetaseries/_version.py
-versionfile_build = src/nibetaseries/_version.py
+versionfile_build = nibetaseries/_version.py
 tag_prefix = v
