### 2019-09-03 04eda60
[MAINT] Change Installation Method (#187)
* create long_description dynamically

* move install to cfg/toml from setup.py

* move dockerfile generator to sub directory

* add cython as dependency (for py3.5?)

* add cython as requirement (for py3.5?)

* fix versionfile_build

* update manifest

* add twine check to make sure long_description does not cause problems
