+++ long_description.rst
@@ -1,0 +1,216 @@
+.. _readme:
+
+============
+NiBetaSeries
+============
+
+If you are viewing this file on GitHub, please see our
+`readthedocs page <https://nibetaseries.readthedocs.io>`_
+for links to render properly.
+
+
+
+What is NiBetaSeries?
+---------------------
+NiBetaSeries is BIDS_-compatible `application <https://bids-apps.neuroimaging.io/>`_
+that calculates betaseries correlations.
+In brief, a beta coefficient (i.e., parameter estimate) is calculated
+for each trial (or event) resulting in a series of betas ("betaseries")
+that can be correlated across regions of interest.
+
+Why should I use it?
+--------------------
+There are potential insights hidden in your task fMRI data.
+Rest fMRI enjoys a multitude of toolboxes which can be applied to task fMRI
+with some effort, but there are not many toolboxes that focus on making
+betaseries.
+Betaseries can then be used for correlations/classifications and
+a multitude of other analyses.
+While a couple alternatives exist (pybetaseries_ and BASCO_), NiBetaSeries
+is the only application to interface with BIDS_ organized data with the goal
+of providing a command-line application experience like fMRIPrep_.
+
+What does NiBetaSeries give me?
+-------------------------------
+Currently NiBetaSeries returns symmetric z-transformed correlation
+matrices, with an entry for each parcel defined in the atlas.
+Soon, NiBetaSeries will also return the betaseries images themselves,
+so you can flexibly apply additional analysis methods.
+
+.. note:: The betas (i.e., parameter estimates) are generated using the "Least Squares Separate" procedure.
+    Please read the betaseries page for more background information.
+    There are plans to support Least Squares All in future iterations.
+
+What do I need to run NiBetaSeries?
+-----------------------------------
+NiBetaSeries takes BIDS_ and preprocessed data as input that satisfy the
+`BIDS derivatives specification <http://bit.ly/2vKeKcp>`_.
+In practical terms, NiBetaSeries uses the output of fMRIPrep_,
+a great BIDS-compatible preprocessing tool.
+NiBetaSeries requires the input and the atlas to already
+be in the same space (e.g., MNI space).
+For more details, see usage and the tutorial
+(sphx_glr_auto_examples_plot_run_nibetaseries.py)
+
+Get Involved
+------------
+This is a very young project that still needs some tender loving care to grow.
+That's where you fit in!
+If you would like to contribute, please read our code_of_conduct
+and contributing page (contributing).
+
+Thanks!
+-------
+This project heavily leverages `nipype <http://nipype.readthedocs.io/en/latest/>`_,
+`nilearn <https://nilearn.github.io/>`_, `pybids <https://bids-standard.github.io/pybids/>`_, and
+`nistats <https://nistats.github.io/>`_ for development.
+Please check out their pages and support the developers.
+
+.. _BASCO: https://www.nitrc.org/projects/basco/
+.. _pybetaseries: https://github.com/poldrack/pybetaseries
+.. _BIDS: http://bids.neuroimaging.io/
+.. _fMRIPrep: http://fmriprep.readthedocs.io/en/latest/
+
+.. _changelog:
+
+0.3.0 (August 29, 2019)
+=======================
+
+Thanks to @PeerHerholz and @njvack for their contributions on this release.
[…]
