+++ scripts/README.md
@@ -1,0 +1,12 @@
+# Misc. Scripts
+
+These scripts help with various tasks for maintaining NiBetaSeries.
+CALL THE SCRIPTS FROM THE BASE OF THE REPOSITORY,
+NOT FROM WITHIN THE SCRIPTS DIRECTORY.
+
+- `generate_nibetaseries_images.sh`
+  - creates a Dockerfile to build a NiBetaSeries Docker image.
+
+- `create_long_description.py`
+  - creates a long description that is valid rst without references to
+    to other files.

[…]
