### 2019-09-04 280eb41
[DOC] address review comments (#185)
* address @labarba review comments

* add (`4D`) to first mention
