### 2019-09-04 9787817
[DOC] capitalize pybetaseries
