### 2019-09-04 b99dcd0
[DOC] betaseries -> beta series
