### 2019-09-04 f498a4c
Update CHANGELOG.rst
