### 2019-09-04 #1295 @jdkent
- The final version number is `v0.3.2`
- here is the doi for the release: 10.5281/zenodo.3385339.
- Only Peer and I are authors, but I did include everyone as "contributors". Is that okay?
