### 2019-09-04 2dd7997
bump pybids version to 0.9.3
