+++ setup.cfg
@@ -29,7 +29,7 @@ keywords =
 python_requires = >=3.5
 install_requires = 
     nipype ~= 1.1.5
-    pybids ~= 0.6.4
+    pybids ~= 0.9.3
     nibabel ~= 2.3.0
     nistats ~= 0.0.1b
     nilearn ~= 0.4.2
