### 2019-09-04 df79d2f
update tests with valid names
