+++ src/nibetaseries/conftest.py
@@ -6,9 +6,75 @@ from nistats.hemodynamic_models import spm_hrf
 from scipy.stats import pearsonr
 from scipy.optimize import minimize
 import pytest
+from bids.layout.writing import build_path
+import pkg_resources
+import os.path as op
 
-bids_prefix = "sub-01_ses-pre_task-waffles"
-deriv_prefix = bids_prefix + "_space-MNI152NLin2009cAsym"
+# read in filename patterns to create valid filenames
+with pkg_resources.resource_stream("bids",
+                                   op.join("layout", "config", "bids.json")) as bids_cfg:
+    bids_patterns = json.load(bids_cfg)['default_path_patterns']
+
+with pkg_resources.resource_stream("nibetaseries",
+                                   op.join("data", "derivatives.json")) as deriv_cfg:
+    deriv_patterns = json.load(deriv_cfg)['fmriprep_path_patterns']
+
+# building blocks for generating valid file names
+subject_entities = {
+    "subject": "01",
+    "session": "pre",
+}
+
+bids_bold_entities = {
+    **subject_entities,
+    "task": "waffles",
+    "suffix": "bold",
+    "extension": "nii.gz",
+}
+bids_bold_fname = build_path(bids_bold_entities, bids_patterns)
+
+bids_json_entities = {
+    **bids_bold_entities,
+    "extension": "json",
+}
+bids_json_fname = build_path(bids_json_entities, bids_patterns)
+
+bids_events_entities = {
+    **bids_bold_entities,
+    "suffix": "events",
+    "extension": "tsv",
+}
+bids_events_fname = build_path(bids_events_entities, bids_patterns)
+
+deriv_bold_entities = {
+    **bids_bold_entities,
+    "space": "MNI152NLin2009cAsym",
+    "description": "preproc",
+}
+deriv_bold_fname = build_path(deriv_bold_entities, deriv_patterns)
+
+deriv_mask_entities = {
+    **deriv_bold_entities,
+    "suffix": "mask",
+    "description": "brain",
+}
+deriv_mask_fname = build_path(deriv_mask_entities, deriv_patterns)
+
+deriv_regressor_entities = {
+    **subject_entities,
+    "task": "waffles",
+    "description": "confounds",
+    "suffix": "regressors",
+    "extension": "tsv",
+}
+deriv_regressor_fname = build_path(deriv_regressor_entities, deriv_patterns)
+
+deriv_betaseries_entities = {
+    **deriv_bold_entities,
+    "suffix": "betaseries",
+    "description": "testCond",
+}
+deriv_betaseries_fname = build_path(deriv_betaseries_entities, deriv_patterns)
 
 
 @pytest.fixture(scope='session')
@@ -17,26 +83,31 @@ def bids_dir(tmpdir_factory):
 
 
 @pytest.fixture(scope='session')
-def sub_bids(bids_dir):
-    return bids_dir.ensure('sub-01',
-                           'ses-pre',
-                           'func',
+def sub_bids(bids_dir, example_file=bids_bold_fname):
+    sub_dir = op.dirname(example_file)
+
+    return bids_dir.ensure(sub_dir,
                            dir=True)
 
 
 @pytest.fixture(scope='session')
-def fmriprep_dir(bids_dir):
+def deriv_dir(bids_dir):
     return bids_dir.ensure('derivatives',
                            'fmriprep',
-                           'sub-01',
-                           'ses-pre',
-                           'func',
[…]
