### 2019-09-04 0f44958
Update documentation regarding version/description of LSS.
