+++ docs/betaseries.rst
@@ -1,23 +1,23 @@
 .. _betaseries:
 
-==========
-Betaseries
-==========
+===========
+Beta series
+===========
 
-This covers the conceptual background of betaseries analysis;
+This covers the conceptual background of beta series analysis;
 check out :ref:`usage` for more on how to use NiBetaSeries.
 
 Introduction
 ------------
-Betaseries track the event-to-event hemodynamic fluctuations
+Beta series track the event-to-event hemodynamic fluctuations
 modeled in task functional magnetic resonance imaging (task fMRI).
-Betaseries fills an important analytical gap between measuring hemodynamic
+Beta series fills an important analytical gap between measuring hemodynamic
 fluctuations in resting-state fMRI and measuring regional activations
 via cognitive subtraction in task fMRI.
 
 Conceptual Background
 ---------------------
-Jesse Rissman :cite:`d-Rissman2004` was the first to publish on betaseries
+Jesse Rissman :cite:`d-Rissman2004` was the first to publish on beta series
 correlations, describing their usage in a working memory task.
 In this task, participants saw a cue, a delay, and a probe, all occurring
 within a short time period.
@@ -29,7 +29,7 @@ takes approximately six seconds to reach its peak, and generally takes over
 The events within the events occur too close to each other to discern what
 brain responses are related to encoding the cue, the delay, or the probe.
 To discern how the activated brain regions form networks, Rissman
-computed betaseries correlations.
+computed beta series correlations.
 Instead of having a single regressor to describe all the cue events,
 a single regressor for all the delay events, and a single regressor for all the
 probe events (as is done in traditional task analysis),
@@ -41,23 +41,31 @@ Once you calculate a beta estimate for each event of a given type
 (e.g., cue), you will have a four-dimensional dataset where each volume
 represents the beta estimates for a particular event.
 
-Having one regressor per event in a single model is known as "least squares all".
-This method, however, has limitations in the context of fast event related
+Having one regressor per event in a single model is known as "least squares all" (LSA).
+This method, however, has limitations in the context of fast event-related
 designs (e.g., designs where the events occur between 3-6
 seconds apart on average).
 Since each event has its own regressor, events that occur very close in time
-are colinear (e.g. are very overlapping).
-Benjamin Turner :cite:`d-Turner2012a` derived a solution for
-the high colinearity observed in least sqaures all by using another
-type of regression known as "least squares separate".
+are collinear (e.g. are very overlapping).
+
+Jeanette Mumford :cit:`d-Mumford2012` derived a solution for
+the high collinearity observed in least squares all by using another
+type of regression known as "least squares separate" (LSS).
 Instead of having one general linear model (GLM) with a regressor per event,
 least squares separate implements a GLM per event with only two regressors:
 1) one for the event of interest, and 2) one for every other event in the
 experiment.
-This process reduces the colinearity of the regressors and creates a more valid
-estimate of how each regressor fits the data.
-NiBetaSeries uses "least squares separate" and is thus optimized
-for fast event related designs.
+This process reduces the collinearity of the regressors and creates a more valid
+estimate for each trial, but also combines all other conditions
+within the design matrix, which will reduce model fit.
+
+Benjamin Turner :cite:`d-Turner2012a` improved upon the LSS method by retaining
+the original conditions in the design matrix.
+In this updated version, the individual trial's design matrix is almost the same
+as the original design matrix, except that the trial is separated out into its
+own regressor.
+NiBetaSeries uses this updated "least squares separate" method and is thus optimized
+for fast event-related designs.
 
 
 Mathematical Background
@@ -74,7 +82,7 @@ The above equation is the general linear model (GLM) presented using
 matrix notation.
 :math:`Y` represents the time-series we are attempting to explain
 (for a given voxel).
[…]
