+++ setup.cfg
@@ -44,6 +44,7 @@ install_requires =
 packages_dir =
     = src
 packages = find:
+include_package_data = True
 
 [options.packages.find]
 where=src
@@ -77,6 +78,10 @@ all =
     %(nb)s
     %(doc)s
     %(dev)s
+
+[options.package_data]
+nibetaseries =
+    data/*json
 
 [options.entry_points]
 console_scripts =
