### 2019-09-05 faf9ac1
refactor bids interface
