### 2019-09-05 38a816a
rename betaseries file output
