### 2019-09-05 1f973b5
make filenames more systematic and bidsy
