### 2019-09-05 3bea70e
- refactor layout wrt 0.9.3 - change variant-> description - add betaseries to output node - update datasink references
