+++ src/nibetaseries/workflows/base.py
@@ -19,9 +19,9 @@ from bids import BIDSLayout
 
 def init_nibetaseries_participant_wf(
     atlas_img, atlas_lut, bids_dir,
-    derivatives_pipeline_dir, exclude_variant_label, hrf_model, low_pass,
+    derivatives_pipeline_dir, exclude_description_label, hrf_model, low_pass,
     output_dir, run_label, selected_confounds, session_label, smoothing_kernel,
-    space_label, subject_list, task_label, variant_label, work_dir,
+    space_label, subject_list, task_label, description_label, work_dir,
         ):
 
     """
@@ -39,8 +39,8 @@ def init_nibetaseries_participant_wf(
             Root directory of BIDS dataset
         derivatives_pipeline_dir: str
             Root directory of the derivatives pipeline
-        exclude_variant_label: str or None
-            Exclude bold series containing this variant label
+        exclude_description_label: str or None
+            Exclude bold series containing this description label
         hrf_model : str
             The model that represents the shape of the hemodynamic response function
         low_pass : float or None
@@ -61,8 +61,8 @@ def init_nibetaseries_participant_wf(
             List of subject labels
         task_label : str or None
             Include bold series containing this task label
-        variant_label : str or None
-            Include bold series containing this variant label
+        description_label : str or None
+            Include bold series containing this description label
         work_dir : str
             Directory in which to store workflow execution state and temporary files
     """
@@ -72,9 +72,7 @@ def init_nibetaseries_participant_wf(
     os.makedirs(nibetaseries_participant_wf.base_dir, exist_ok=True)
 
     # reading in derivatives and bids inputs as queryable database like objects
-    layout = BIDSLayout([(bids_dir, ['bids']),
-                         (derivatives_pipeline_dir, ['bids', 'derivatives'])],
-                        include=['sub-*', 'dataset_description.json', 'task-*'])
+    layout = BIDSLayout(bids_dir, derivatives=derivatives_pipeline_dir)
 
     for subject_label in subject_list:
 
@@ -85,7 +83,7 @@ def init_nibetaseries_participant_wf(
                                     run=run_label,
                                     ses=session_label,
                                     space=space_label,
-                                    variant=variant_label)
+                                    description=description_label)
         # collect files to be associated with each preproc
         brainmask_list = [d['brainmask'] for d in subject_data]
         confound_tsv_list = [d['confounds'] for d in subject_data]
@@ -110,7 +108,7 @@ def init_nibetaseries_participant_wf(
         )
 
         single_subject_wf.config['execution']['crashdump_dir'] = (
-            os.path.join(output_dir, "nibetaseries", "sub-" + subject_label, 'log')
+            os.path.join(output_dir, "sub-" + subject_label, 'log')
         )
 
         for node in single_subject_wf._get_all_nodes():
@@ -229,7 +227,8 @@ def init_single_subject_wf(
     input_node.inputs.atlas_lut = atlas_lut
 
     output_node = pe.Node(niu.IdentityInterface(fields=['correlation_matrix',
-                                                        'correlation_fig']),
+                                                        'correlation_fig',
+                                                        'betaseries_file']),
                           name='output_node')
 
     # initialize the betaseries workflow
@@ -241,15 +240,17 @@ def init_single_subject_wf(
     correlation_wf = init_correlation_wf()
 
     # correlation matrix datasink
-    ds_correlation_matrix = pe.MapNode(DerivativesDataSink(base_directory=output_dir,
-                                                           suffix='matrix'),
-                                       iterfield=['betaseries_file', 'in_file'],
[…]
