### 2019-09-05 5d23af3
fix grammer (e.g. ..) -> (e.g., ...)
