### 2019-09-05 270661c
change variant -> desc in usage documentation
