### 2019-09-05 4ca7c49
refactor/add tests wrt pybids 0.9.3
