### 2019-09-05 8083398
change how to read in local json files
