+++ src/nibetaseries/interfaces/bids.py
@@ -40,9 +40,8 @@ class DerivativesDataSink(SimpleInterface):
         from bids.layout import parse_file_entities
         from bids.layout.writing import build_path
 
-        with pkg_resources.resource_stream("nibetaseries",
-                                           op.join("data", "derivatives.json")) as deriv_cfg:
-            deriv_patterns = json.load(deriv_cfg)['fmriprep_path_patterns']
+        deriv_cfg = pkg_resources.resource_string("nibetaseries", op.join("data", "derivatives.json"))
+        deriv_patterns = json.loads(deriv_cfg.decode('utf-8'))['fmriprep_path_patterns']
 
         subject_entities = parse_file_entities(self.inputs.source_file)
         betaseries_entities = parse_file_entities(self.inputs.in_file)
