### 2019-09-05 f2f0d26
pacify flake8 for bids.py
