### 2019-09-05 4488e2e
coerce sub_metadata as a string
