### 2019-09-05 f0df0d8
fix documentation/update docs for beta series images being passed out
