### 2019-09-05 0c507a5
[ENH] Separate other conditions in LSS model (#191)
* Retain conditions' trial types.

* It was because of the underscore in the trial type I think.

* Update documentation regarding version/description of LSS.

* Fix in-text citation.

* - update code example
  - fix some grammar
  - make headings consistently spaced

* fix code typo [skip ci]
