### 2019-09-05 1ca2aca
[ENH,FIX] refactor bids file processing (#193)
* bump pybids version to 0.9.3

* update tests with valid names

* add patterns for building bids files

* add data files to package

* switch variant -> desc and remove extraneous directory

* refactor bids interface

* rename betaseries file output

* make filenames more systematic and bidsy

* refactor data collection wrt pybids 0.9.3

* remove _rename_matrix from analysis.py

* - refactor layout wrt 0.9.3
  - change variant-> description
  - add betaseries to output node
  - update datasink references

* fix grammer (e.g. ..) -> (e.g., ...)

*
[…]
