+++ src/nibetaseries/workflows/analysis.py
@@ -3,7 +3,6 @@
 # emacs: -*- mode: python; py-indent-offset: 4; indent-tabs-mode: nil -*-
 # vi: set ft=python sts=4 ts=4 sw=4 et:
 
-# TODO: fix the renaming of files
 """
 This workflow takes roi-roi correlations of the input betaseries
 """
@@ -67,35 +66,12 @@ def init_correlation_wf(name="correlation_wf"):
                                  name='atlas_corr_node',
                                  iterfield=['timeseries_file'])
 
-    rename_matrix_node = pe.MapNode(niu.Function(output_names=['correlation_matrix_trialtype'],
-                                                 function=_rename_matrix),
-                                    iterfield=['correlation_matrix', 'betaseries_file'],
-                                    name='rename_matrix_node')
     workflow.connect([
         (input_node, atlas_corr_node, [('betaseries_files', 'timeseries_file'),
                                        ('atlas_file', 'atlas_file'),
                                        ('atlas_lut', 'atlas_lut')]),
-        (input_node, rename_matrix_node, [('betaseries_files', 'betaseries_file')]),
-        (atlas_corr_node, rename_matrix_node, [('correlation_matrix', 'correlation_matrix')]),
-        (rename_matrix_node, output_node, [('correlation_matrix_trialtype',
-                                            'correlation_matrix')]),
-        (atlas_corr_node, output_node, [('correlation_fig', 'correlation_fig')])
+        (atlas_corr_node, output_node, [('correlation_fig', 'correlation_fig'),
+                                        ('correlation_matrix', 'correlation_matrix')])
     ])
 
     return workflow
-
-
-def _rename_matrix(correlation_matrix, betaseries_file):
-    import os
-    import re
-    from shutil import copyfile
-    # import pdb; pdb.set_trace()
-    betaseries_regex = re.compile('.*betaseries_trialtype-(?P<trial_type>[A-Za-z0-9_]+).nii.gz')
-    trial_type = betaseries_regex.search(
-        betaseries_file).groupdict()['trial_type'].replace('_', '')
-    out_file = os.path.join(os.getcwd(),
-                            'correlation-matrix_trialtype-{trial_type}.tsv'.format(
-                                trial_type=trial_type))
-    copyfile(correlation_matrix, out_file)
-
-    return out_file
