### 2019-09-09 b608fde
Switch Nipype workflows to LiterateWorkflows for boilerplate.
Also add a couple of references to references.bib.
