+++ src/nibetaseries/workflows/base.py
@@ -14,6 +14,8 @@ from .analysis import init_correlation_wf
 from ..interfaces.bids import DerivativesDataSink
 from nipype.pipeline import engine as pe
 from nipype.interfaces import utility as niu
+from nipype import __version__ as nipype_ver
+from niworkflows.engine.workflows import LiterateWorkflow as Workflow
 from bids import BIDSLayout
 
 
@@ -67,9 +69,27 @@ def init_nibetaseries_participant_wf(
             Directory in which to store workflow execution state and temporary files
     """
     # setup workflow
-    nibetaseries_participant_wf = pe.Workflow(name='nibetaseries_participant_wf')
+    nibetaseries_participant_wf = Workflow(name='nibetaseries_participant_wf')
     nibetaseries_participant_wf.base_dir = os.path.join(work_dir, 'NiBetaSeries_work')
     os.makedirs(nibetaseries_participant_wf.base_dir, exist_ok=True)
+
+    workflow.__desc__ = """
+Results included in this manuscript come from modeling
+performed using *NiBetaSeries* {nbs_ver} (@NiBetaSeriesZenodo),
+which is based on *Nipype* {nipype_ver} (@nipype1; @nipype2; RRID:SCR_002502).
+""".format(nbs_ver='1.0.0', nipype_ver=nipype_ver)
+
+    workflow.__postdesc__ = """
+
+### Copyright Waiver
+The above boilerplate text was automatically generated by NiBetaSeries
+with the express intention that users should copy and paste this
+text into their manuscripts *unchanged*.
+It is released under the [CC0]\
+(https://creativecommons.org/publicdomain/zero/1.0/) license.
+
+### References
+""".format(nilearn_ver=nilearn_ver)
 
     # reading in derivatives and bids inputs as queryable database like objects
     layout = BIDSLayout(bids_dir, derivatives=derivatives_pipeline_dir)
