### 2019-09-09 f366480
Replace low-pass filter with high-pass filter.
Also fix a couple of typos in docs/workflows.rst.
