+++ src/nibetaseries/cli/run.py
@@ -64,8 +64,8 @@ def get_parser():
     proc_opts = parser.add_argument_group('Options for processing')
     proc_opts.add_argument('-sm', '--smoothing-kernel', action='store', type=float, default=6.0,
                            help='select a smoothing kernel (mm)')
-    proc_opts.add_argument('-lp', '--low-pass', action='store', type=float,
-                           default=None, help='low pass filter (Hz)')
+    proc_opts.add_argument('-hp', '--high-pass', action='store', type=float,
+                           default=None, help='high pass filter (Hz)')
     proc_opts.add_argument('-c', '--confounds', help='The confound column names '
                            'that are to be included in nuisance regression. '
                            'write the confounds you wish to include separated by a space',
@@ -201,7 +201,7 @@ def main():
             derivatives_pipeline_dir=derivatives_pipeline_dir,
             exclude_description_label=opts.exclude_description_label,
             hrf_model=opts.hrf_model,
-            low_pass=opts.low_pass,
+            high_pass=opts.high_pass,
             output_dir=output_dir,
             run_label=opts.run_label,
             selected_confounds=opts.confounds,
