### 2019-09-09 bf52f29
explicitly set codecov settings
