### 2019-09-09 0ab3109
Change HPF of 0.01 to 0.008 in tests.
Much closer to 128, and was apparently causing the errors.
