+++ src/nibetaseries/cli/tests/test_run.py
@@ -17,7 +17,7 @@ def test_nibs(bids_dir, atlas_file, atlas_lut, deriv_dir):
                  "-a " + str(atlas_file),
                  "-l " + str(atlas_lut),
                  "-c WhiteMatter CSF",
-                 "-hp 0.01",
+                 "-hp 0.008",
                  bids_dir,
                  "fmriprep",
                  out_dir,
