### 2019-09-09 #1295 @arokem
@openjournals/joss-eics : is everything in order here?
