+++ src/nibetaseries/workflows/base.py
@@ -73,13 +73,13 @@ def init_nibetaseries_participant_wf(
     nibetaseries_participant_wf.base_dir = os.path.join(work_dir, 'NiBetaSeries_work')
     os.makedirs(nibetaseries_participant_wf.base_dir, exist_ok=True)
 
-    workflow.__desc__ = """
+    nibetaseries_participant_wf.__desc__ = """
 Results included in this manuscript come from modeling
 performed using *NiBetaSeries* {nbs_ver} (@NiBetaSeriesZenodo),
 which is based on *Nipype* {nipype_ver} (@nipype1; @nipype2; RRID:SCR_002502).
 """.format(nbs_ver='1.0.0', nipype_ver=nipype_ver)
 
-    workflow.__postdesc__ = """
+    nibetaseries_participant_wf.__postdesc__ = """
 
 ### Copyright Waiver
 The above boilerplate text was automatically generated by NiBetaSeries
@@ -89,13 +89,12 @@ It is released under the [CC0]\
 (https://creativecommons.org/publicdomain/zero/1.0/) license.
 
 ### References
-""".format(nilearn_ver=nilearn_ver)
+"""
 
     # reading in derivatives and bids inputs as queryable database like objects
     layout = BIDSLayout(bids_dir, derivatives=derivatives_pipeline_dir)
 
     for subject_label in subject_list:
-
         # collect the necessary inputs for both collect data
         subject_data = collect_data(layout,
                                     subject_label,
@@ -224,7 +223,7 @@ def init_single_subject_wf(
         correlation_fig
             a svg file of a circular connectivity plot showing all roi-roi correlations
     """
-    workflow = pe.Workflow(name=name)
+    workflow = Workflow(name=name)
 
     # name the nodes
     input_node = pe.Node(niu.IdentityInterface(fields=['atlas_img',
