### 2019-09-09 dbd3978
Consolidate BibTeX files.
