+++ docs/references.bib
@@ -1,159 +1,0 @@
-@article{Gottlich2015,
-abstract = {BASCO (BetA Series COrrelation) is a user-friendly MATLAB toolbox with a graphical user interface (GUI) which allows investigating functional connectivity in event-related functional magnetic resonance imaging (fMRI) data. Connectivity analyses extend and compliment univariate activation analyses since the actual interaction between brain regions involved in a task can be explored. BASCO supports seed-based functional connectivity as well as brain network analyses. Although there are a multitude of advanced toolboxes for investigating resting-state functional connectivity, BASCO is the first toolbox for evaluating task-related whole-brain functional connectivity employing a large number of network nodes. Thus, BASCO allows investigating task-specific rather than resting-state networks. Here, we summarize the main features of the toolbox and describe the methods and algorithms.},
-author = {G{\"{o}}ttlich, Martin and Beyer, Frederike and Kr{\"{a}}mer, Ulrike M.},
-doi = {10.3389/fnsys.2015.00126},
-isbn = {1662-5137 (Electronic)$\backslash$r1662-5137 (Linking)},
-issn = {1662-5137},
-journal = {Frontiers in Systems Neuroscience},
-keywords = {b,beta-series correlation,betaseries,degree centrality,event-related design,fMRI,fmri,functional connectivity,method,network analysis},
-mendeley-tags = {betaseries,method},
-number = {September},
-pages = {1--10},
-pmid = {26441558},
-title = {{BASCO: a toolbox for task-related functional connectivity}},
-url = {http://journal.frontiersin.org/Article/10.3389/fnsys.2015.00126/abstract},
-volume = {9},
-year = {2015}
-}
-@article{Esteban2019,
-author = {Esteban, Oscar and Markiewicz, Christopher J and Blair, Ross W and Moodie, Craig A and Ayse, I and Erramuzpe, Asier and Kent, James D and Goncalves, Mathias and Dupre, Elizabeth and Snyder, Madeleine and Oya, Hiroyuki and Ghosh, Satrajit S and Wright, Jessey and Durnez, Joke and Poldrack, Russell A and Gorgolewski, Krzysztof J},
-doi = {10.1038/s41592-018-0235-4},
-file = {:media/Data/Downloads/fmriprep.pdf:pdf},
-journal = {Nature methods},
-number = {1},
-pages = {111--116},
-title = {{FMRIPrep : a robust preprocessing pipeline for functional MRI}},
-volume = {16},
-year = {2019}
-}
-@article{Rissman2004,
[…]
