+++ docs/betaseries.rst
@@ -300,7 +300,7 @@ Other Relevant Readings
 References
 ----------
 
-.. bibliography:: references.bib
+.. bibliography:: ../src/nibetaseries/data/references.bib
     :style: plain
     :labelprefix: docs-
     :keyprefix: d-
