### 2019-09-09 389e3c5
make doc strings consistent
