### 2019-09-09 dc41ad2
change high pass default from None -> 0.0078125
