### 2019-09-10 6f38669
[FIX, DOC] Rename low-pass filter to high-pass filter (#198)
* Replace low-pass filter with high-pass filter.
  Also fix a couple of typos in docs/workflows.rst.

* Add tsalo to Zenodo file.

* Fix spacing.

* Change HPF default.

* Fix tests and docs examples.

* Change HPF of 0.01 to 0.008 in tests.

  Much closer to 128, and was apparently causing the errors.

* make doc strings consistent

* change high pass default from None -> 0.0078125
