### 2019-09-10 50e0e62
test more conditions in utils.py
