### 2019-09-10 bc262f3
[TST] add more tests (#201)
* add rest file and add a run number

* more comprehensively test run.py

* add run information

* test more conditions in utils.py
