### 2019-09-10 1e503ef
Implement LSA and add estimator argument.
