+++ src/nibetaseries/interfaces/nistats.py
@@ -14,7 +14,7 @@ class NistatsBaseInterface(LibraryBaseInterface):
     _pkg = 'nistats'
 
 
-class BetaSeriesInputSpec(BaseInterfaceInputSpec):
+class LSSBetaSeriesInputSpec(BaseInterfaceInputSpec):
     bold_file = File(exists=True, mandatory=True,
                      desc="The bold run")
     bold_metadata = traits.Dict(desc='Dictionary containing useful information about'
@@ -33,15 +33,15 @@ class BetaSeriesInputSpec(BaseInterfaceInputSpec):
     high_pass = traits.Float(0.0078125, desc="the high pass filter (Hz)")
 
 
-class BetaSeriesOutputSpec(TraitedSpec):
+class LSSBetaSeriesOutputSpec(TraitedSpec):
     beta_maps = OutputMultiPath(File)
     design_matrices = traits.Dict()
 
 
-class BetaSeries(NistatsBaseInterface, SimpleInterface):
+class LSSBetaSeries(NistatsBaseInterface, SimpleInterface):
     """Calculates BetaSeries Maps From a BOLD file (one series map per event type)."""
-    input_spec = BetaSeriesInputSpec
-    output_spec = BetaSeriesOutputSpec
+    input_spec = LSSBetaSeriesInputSpec
+    output_spec = LSSBetaSeriesOutputSpec
 
     def _run_interface(self, runtime):
         from nistats import first_level_model
@@ -62,16 +62,18 @@ class BetaSeries(NistatsBaseInterface, SimpleInterface):
         high_pass_period = int(1 / self.inputs.high_pass)
 
         # setup the model
-        model = first_level_model.FirstLevelModel(t_r=t_r,
-                                                  slice_time_ref=0,
-                                                  hrf_model=self.inputs.hrf_model,
-                                                  mask=self.inputs.mask_file,
-                                                  smoothing_fwhm=self.inputs.smoothing_kernel,
-                                                  standardize=True,
-                                                  signal_scaling=0,
-                                                  period_cut=high_pass_period,
-                                                  drift_model='cosine',
-                                                  verbose=1)
+        model = first_level_model.FirstLevelModel(
+            t_r=t_r,
+            slice_time_ref=0,
+            hrf_model=self.inputs.hrf_model,
+            mask=self.inputs.mask_file,
+            smoothing_fwhm=self.inputs.smoothing_kernel,
+            standardize=True,
+            signal_scaling=0,
+            period_cut=high_pass_period,
+            drift_model='cosine',
+            verbose=1
+        )
 
         # initialize dictionary to contain trial estimates (betas)
         beta_maps = {}
@@ -111,9 +113,137 @@ class BetaSeries(NistatsBaseInterface, SimpleInterface):
                 nib.save(beta_series, beta_series_template.format(trial_type=t_type))
                 beta_series_lst.append(beta_series_template.format(trial_type=t_type))
 
-            self._results['beta_maps'] = beta_series_lst
-            self._results['design_matrices'] = design_matrix_collector
+        self._results['beta_maps'] = beta_series_lst
+        self._results['design_matrices'] = design_matrix_collector
         return runtime
+
+
+class LSABetaSeriesInputSpec(BaseInterfaceInputSpec):
+    bold_file = File(exists=True, mandatory=True,
+                     desc="The bold run")
+    bold_metadata = traits.Dict(desc='Dictionary containing useful information about'
+                                ' the bold_file')
+    mask_file = File(exists=True, mandatory=True,
+                     desc="Binarized nifti file indicating the brain")
+    events_file = File(exists=True, mandatory=True,
+                       desc="File that contains all events from the bold run")
+    confounds_file = traits.Either(None, File(exists=True),
[…]
