+++ src/nibetaseries/interfaces/tests/test_nistats.py
@@ -5,14 +5,39 @@ import json
 from ..nistats import BetaSeries
 
 
-def test_beta_series(sub_metadata, preproc_file, sub_events,
-                     confounds_file, brainmask_file):
+def test_lss_beta_series(sub_metadata, preproc_file, sub_events,
+                         confounds_file, brainmask_file):
     selected_confounds = ['WhiteMatter', 'CSF']
     hrf_model = 'spm'
     with open(str(sub_metadata), 'r') as md:
         bold_metadata = json.load(md)
 
-    beta_series = BetaSeries(bold_file=str(preproc_file),
+    beta_series = BetaSeries(estimator='lss',
+                             bold_file=str(preproc_file),
+                             bold_metadata=bold_metadata,
+                             mask_file=str(brainmask_file),
+                             events_file=str(sub_events),
+                             confounds_file=str(confounds_file),
+                             selected_confounds=selected_confounds,
+                             hrf_model=hrf_model,
+                             smoothing_kernel=None,
+                             high_pass=0.008)
+    res = beta_series.run()
+
+    assert os.path.isfile(res.outputs.beta_maps)
+
+    os.remove(res.outputs.beta_maps)
+
+
+def test_lsa_beta_series(sub_metadata, preproc_file, sub_events,
+                         confounds_file, brainmask_file):
+    selected_confounds = ['WhiteMatter', 'CSF']
+    hrf_model = 'spm'
+    with open(str(sub_metadata), 'r') as md:
+        bold_metadata = json.load(md)
+
+    beta_series = BetaSeries(estimator='lsa',
+                             bold_file=str(preproc_file),
                              bold_metadata=bold_metadata,
                              mask_file=str(brainmask_file),
                              events_file=str(sub_events),
