### 2019-09-10 f67cdca
Revert example back to LSS.
