### 2019-09-10 #1295 @labarba
I'm not on rotation, but could you now check that the issues I flagged were resolved (version, etc.), and update as needed with `whedon`?
