### 2019-09-11 2d96ccb
say we support LSA in readme
