+++ src/nibetaseries/interfaces/nistats.py
@@ -89,7 +89,7 @@ class LSSBetaSeries(NistatsBaseInterface, SimpleInterface):
             # calculate the beta map
             beta_map = model.compute_contrast(trial_type, output_type='effect_size')
             design_matrix_collector[trial_idx] = model.design_matrices_[0]
-            # import pdb; pdb.set_trace()
+
             # assign beta map to appropriate list
             if trial_type in beta_maps:
                 beta_maps[trial_type].append(beta_map)
@@ -190,7 +190,7 @@ class LSABetaSeries(NistatsBaseInterface, SimpleInterface):
 
             # calculate the beta map
             beta_map = model.compute_contrast(t_name, output_type='effect_size')
-            # import pdb; pdb.set_trace()
+
             # assign beta map to appropriate list
             if t_type in beta_maps:
                 beta_maps[t_type].append(beta_map)
