### 2019-09-11 82c0d06
add additional tests
