+++ src/nibetaseries/interfaces/tests/test_nistats.py
@@ -3,6 +3,7 @@ import os
 import json
 
 from ..nistats import LSSBetaSeries, LSABetaSeries
+from ..nistats import _lss_events_iterator, _lsa_events_converter
 
 
 def test_lss_beta_series(sub_metadata, preproc_file, sub_events,
@@ -23,9 +24,9 @@ def test_lss_beta_series(sub_metadata, preproc_file, sub_events,
                                 high_pass=0.008)
     res = beta_series.run()
 
-    assert os.path.isfile(res.outputs.beta_maps)
-
-    os.remove(res.outputs.beta_maps)
+    for beta_map in res.outputs.beta_maps:
+        assert os.path.isfile(beta_map)
+        os.remove(beta_map)
 
 
 def test_lsa_beta_series(sub_metadata, preproc_file, sub_events,
@@ -46,6 +47,30 @@ def test_lsa_beta_series(sub_metadata, preproc_file, sub_events,
                                 high_pass=0.008)
     res = beta_series.run()
 
-    assert os.path.isfile(res.outputs.beta_maps)
+    for beta_map in res.outputs.beta_maps:
+        assert os.path.isfile(beta_map)
+        os.remove(beta_map)
 
-    os.remove(res.outputs.beta_maps)
+
+def test_lss_events_iterator(sub_events):
+    # all but the first instance of waffle
+    # should be changed to "other"
+    t_lst = ['other', 'fry', 'milkshake'] * 5
+    t_lst[0] = 'waffle'
+    res = _lss_events_iterator(sub_events)
+    out_df = list(res)[0][0]
+    out_lst = list(out_df['trial_type'])
+
+    assert t_lst == out_lst
+
+
+def test_lsa_events_converter(sub_events):
+    # each instance of waffle, fry, and milkshake
+    # should have a different number
+    trial_type_lst = ['waffle', 'fry', 'milkshake'] * 5
+    number_lst = ['000{}'.format(x) for x in range(1, 6) for y in range(0, 3)]
+    t_lst = ['_'.join([trial, num]) for trial, num in zip(trial_type_lst, number_lst)]
+    res = _lsa_events_converter(sub_events)
+    out_lst = list(res['trial_type'])
+
+    assert t_lst == out_lst
