### 2019-09-11 9c2a841
fix new error from more complex events file - some pearson correlations were above 1.0
