### 2019-09-11 #1295 @Kevin-Mattheus-Moerman
@jdkent can you work on/check the following for your paper:

- [ ] We are about to process your paper for acceptance. Can you please double check the paper thoroughly and make any final changes (refer to them briefly here if you make any changes). Also ensure the author names and the acknowledgements (if any), and links to funding sources etc, are accurate.
- [ ] Can you check your affiliation. Is "Neuroscience Program" correct? Usually people list a department rather than a program (also it should perhaps be: "Interdisciplinary Graduate Program in Neuroscience"?), should it perhaps be "Psychological and Brain Sciences" instead, like in your [profile page](https://neuroscience.grad.uiowa.edu/content/james-kent)?
- [ ] Can you make the affiliations more complete by adding the city and country as well please?
- [ ] Consider rephrasing the following (which sounds awkward) "Three dimensions are brain voxels,..." to something like "The first three dimensions represent the voxel dimensions,... "
- [ ] Insert missing "be" in this sentence: "For many intents and purposes, the 4D beta series can **[be]** analyzed...."
- [ ] When finished making the above changes please run '@whedon generate pdf' to update your proof
