### 2019-09-11 #1295 @arfon
> @arfon the link to the archive in the PDF does not seem to work. Is this normal at this point?

The link to the archive works in the final proofs (when you do `@whedon accept`)
