+++ src/nibetaseries/workflows/base.py
@@ -19,7 +19,8 @@ from bids import BIDSLayout
 
 def init_nibetaseries_participant_wf(
     atlas_img, atlas_lut, bids_dir,
-    derivatives_pipeline_dir, exclude_description_label, hrf_model, high_pass,
+    derivatives_pipeline_dir, exclude_description_label, fir_delays,
+    hrf_model, high_pass,
     output_dir, run_label, selected_confounds, session_label, smoothing_kernel,
     space_label, subject_list, task_label, description_label, work_dir,
         ):
@@ -37,10 +38,12 @@ def init_nibetaseries_participant_wf(
             Path to input atlas lookup table (tsv)
         bids_dir : str
             Root directory of BIDS dataset
-        derivatives_pipeline_dir: str
+        derivatives_pipeline_dir : str
             Root directory of the derivatives pipeline
-        exclude_description_label: str or None
+        exclude_description_label : str or None
             Exclude bold series containing this description label
+        fir_delays : list or None
+            FIR delays (in scans)
         hrf_model : str
             The model that represents the shape of the hemodynamic response function
         high_pass : float
@@ -99,6 +102,7 @@ def init_nibetaseries_participant_wf(
             brainmask_list=brainmask_list,
             confound_tsv_list=confound_tsv_list,
             events_tsv_list=events_tsv_list,
+            fir_delays=fir_delays,
             hrf_model=hrf_model,
             high_pass=high_pass,
             name='single_subject' + subject_label + '_wf',
@@ -122,8 +126,8 @@ def init_nibetaseries_participant_wf(
 
 def init_single_subject_wf(
     atlas_img, atlas_lut, bold_metadata_list, brainmask_list, confound_tsv_list,
-    events_tsv_list, hrf_model, high_pass, name, output_dir, preproc_img_list,
-    selected_confounds, smoothing_kernel
+    events_tsv_list,  fir_delays, hrf_model, high_pass, name, output_dir,
+    preproc_img_list, selected_confounds, smoothing_kernel
         ):
     """
     This workflow completes the generation of the betaseries files
@@ -140,6 +144,7 @@ def init_single_subject_wf(
             brainmask_list=[''],
             confound_tsv_list=[''],
             events_tsv_list=[''],
+            fir_delays=None,
             hrf_model='',
             high_pass='',
             name='subtest',
@@ -163,6 +168,8 @@ def init_single_subject_wf(
             list of confound tsvs (e.g. from FMRIPREP)
         events_tsv_list : list
             list of event tsvs
+        fir_delays : list or None
+            FIR delays (in scans)
         hrf_model : str
             hemodynamic response function used to model the data
         high_pass : float
@@ -233,7 +240,9 @@ def init_single_subject_wf(
                           name='output_node')
 
     # initialize the betaseries workflow
-    betaseries_wf = init_betaseries_wf(hrf_model=hrf_model, high_pass=high_pass,
+    betaseries_wf = init_betaseries_wf(fir_delays=fir_delays,
+                                       hrf_model=hrf_model,
+                                       high_pass=high_pass,
                                        selected_confounds=selected_confounds,
                                        smoothing_kernel=smoothing_kernel)
 
