### 2019-09-11 #1295 @jdkent
@Kevin-Mattheus-Moerman, thank you for your suggestions! For making the proof, is it alright if the paper on zenodo/GitHub tag is slightly different than the proof that is used in the journal?

I can email zenodo and ask them to update the repository and I can modify which commit `v0.3.2` points to (but I cannot change the pypi release).

However, it would be easier if I could edit the JOSS paper on master and have the git tag and zenodo archive contain a slightly different JOSS paper.
