### 2019-09-11 2b4fec5
Refactor _fisher_r_to_z slightly.
