### 2019-09-11 19655d9
[ENH] Implement least squares- all (LSA) modeling (#202)
* Implement LSA and add estimator argument. @tsalo 

* Update tests. @tsalo 

* Revert example back to LSS. @tsalo 

* Fix bugs. @tsalo 

* Fix docs examples. @tsalo 

* say nibetaseries does both LSS and LSA @jdkent 

* say we support LSA in readme @jdkent 

* remove old pdb imports @jdkent 

* make events file more complex @jdkent 

* add additional tests @jdkent 

* fix new error from more complex events file @jdkent 
  - some pearson correlations were above 1.0

* Refactor _fisher_r_to_z slightly. @tsalo
