### 2019-09-11 HBClab/NiBetaSeries#170 @jdkent
with the advent #193 beta series are now being passed directly to output so people can use the series at the voxel level resolution. (still good to note that the inefficiency for calculating parcels still exists).

closing this issue.
