+++ src/nibetaseries/cli/tests/test_run.py
@@ -19,11 +19,35 @@ def test_nibs_lss(bids_dir, atlas_file, atlas_lut, deriv_dir):
                  "-a " + str(atlas_file),
                  "-l " + str(atlas_lut),
                  "-c WhiteMatter CSF",
-                 "-hp 0.008",
+                 "--high-pass 0.008",
                  "--estimator lss",
                  "-sp MNI152NLin2009cAsym",
                  "-sm 0.0",
                  "--hrf-model spm",
+                 "--session-label pre",
+                 "--task-label waffles",
+                 "--run-label 1",
+                 "--description-label preproc",
+                 "--graph",
+                 bids_dir,
+                 "fmriprep",
+                 out_dir,
+                 "participant"])
+
+
+def test_nibs_fs(bids_dir, atlas_file, atlas_lut, deriv_dir):
+    bids_dir = str(bids_dir)
+    out_dir = os.path.join(bids_dir, 'derivatives')
+    assert call(["nibs",
+                 "-a " + str(atlas_file),
+                 "-l " + str(atlas_lut),
+                 "-c WhiteMatter CSF",
+                 "--high-pass 0.008",
+                 "--estimator lss",
+                 "-sp MNI152NLin2009cAsym",
+                 "-sm 0.0",
+                 "--hrf-model fir",
+                 "--fir-delays 0 1 2 3 4"
                  "--session-label pre",
                  "--task-label waffles",
                  "--run-label 1",
@@ -42,7 +66,7 @@ def test_nibs_lsa(bids_dir, atlas_file, atlas_lut, deriv_dir):
                  "-a " + str(atlas_file),
                  "-l " + str(atlas_lut),
                  "-c WhiteMatter CSF",
-                 "-hp 0.008",
+                 "--high-pass 0.008",
                  "--estimator lsa",
                  "-sp MNI152NLin2009cAsym",
                  "-sm 0.0",
