### 2019-09-13 c096a1e
Add more references and improve boilerplate rendering.
