+++ src/nibetaseries/workflows/base.py
@@ -8,15 +8,23 @@ NiBetaSeries processing workflows
 from __future__ import print_function, division, absolute_import, unicode_literals
 import os
 from copy import deepcopy
+
+from bids import BIDSLayout
+from nipype.pipeline import engine as pe
+from nipype.interfaces import utility as niu
+from niworkflows.engine.workflows import LiterateWorkflow as Workflow
+
+from bids import __version__ as pybids_ver
+from numpy import __version__ as numpy_ver
+from pandas import __version__ as pandas_ver
+from nipype import __version__ as nipype_ver
+from nibabel import __version__ as nibabel_ver
+from niworkflows import __version__ as niworkflows_ver
+
 from .utils import collect_data
 from .model import init_betaseries_wf
 from .analysis import init_correlation_wf
 from ..interfaces.bids import DerivativesDataSink
-from nipype.pipeline import engine as pe
-from nipype.interfaces import utility as niu
-from nipype import __version__ as nipype_ver
-from niworkflows.engine.workflows import LiterateWorkflow as Workflow
-from bids import BIDSLayout
 
 
 def init_nibetaseries_participant_wf(
@@ -76,11 +84,18 @@ def init_nibetaseries_participant_wf(
 
     nibetaseries_participant_wf.__desc__ = """
 Results included in this manuscript come from modeling
-performed using *NiBetaSeries* {nbs_ver} (@NiBetaSeriesZenodo),
-which is based on *Nipype* {nipype_ver} (@nipype1; @nipype2; RRID:SCR_002502).
+performed using *NiBetaSeries* {nbs_ver} [@NiBetaSeriesZenodo],
+which is based on *Nipype* {nipype_ver} [@nipype1; @nipype2].
 """.format(nbs_ver='1.0.0', nipype_ver=nipype_ver)
 
     nibetaseries_participant_wf.__postdesc__ = """
+
+### Software Dependencies
+
+Additional libraries used in the NiBetaSeries workflow include
+*Pybids* {pybids_ver} [@yarkoni2019pybids], *Niworkflows* {niworkflows_ver},
+*Nibabel* {nibabel_ver}, *Pandas* {pandas_ver} [@mckinney2010data], and
+*Numpy* {numpy_ver} [@van2011numpy; @oliphant2006guide].
 
 ### Copyright Waiver
 The above boilerplate text was automatically generated by NiBetaSeries
@@ -90,7 +105,11 @@ It is released under the [CC0]\
 (https://creativecommons.org/publicdomain/zero/1.0/) license.
 
 ### References
-"""
+""".format(pybids_ver=pybids_ver,
+           niworkflows_ver=niworkflows_ver,
+           nibabel_ver=nibabel_ver,
+           pandas_ver=pandas_ver,
+           numpy_ver=numpy_ver)
 
     # reading in derivatives and bids inputs as queryable database like objects
     layout = BIDSLayout(bids_dir, derivatives=derivatives_pipeline_dir)
