### 2019-09-13 b610ef3
Update dependencies in Dockerfile and setup.
