+++ setup.cfg
@@ -20,14 +20,14 @@ classifiers =
     Programming Language :: Python :: 3.6
     Programming Language :: Python :: 3.7
     Topic :: Utilities
-keywords = 
+keywords =
     bids
     fmri
     neuroimaging
 
 [options]
 python_requires = >=3.5
-install_requires = 
+install_requires =
     nipype ~= 1.1.5
     pybids ~= 0.9.3
     nibabel ~= 2.3.0
@@ -35,6 +35,7 @@ install_requires =
     nilearn ~= 0.4.2
     pandas ~= 0.24.0
     numpy
+    niworkflows
     duecredit ~= 0.6.4
     scikit-learn ~= 0.19.2
     matplotlib ~= 2.2.4
@@ -62,7 +63,7 @@ dev =
     flake8
     codecov
     coverage
-doc = 
+doc =
     sphinx >= 1.3
     sphinx_rtd_theme
     sphinx-argparse
