+++ Dockerfile
@@ -1,10 +1,10 @@
 # Generated by Neurodocker version 0.5.0
 # Timestamp: 2019-08-31 17:22:00 UTC
-# 
+#
 # Thank you for using Neurodocker. If you discover any issues
 # or ways to improve this software, please submit an issue or
 # pull request on our GitHub repository:
-# 
+#
 #     https://github.com/kaczmarj/neurodocker
 
 FROM neurodebian:stretch-non-free
@@ -52,6 +52,11 @@ RUN test "$(getent passwd neuro)" || useradd --no-user-group --create-home --she
 USER neuro
 
 WORKDIR /home/neuro
+
+# Install latest pandoc
+RUN curl -o pandoc-2.2.2.1-1-amd64.deb -sSL "https://github.com/jgm/pandoc/releases/download/2.2.2.1/pandoc-2.2.2.1-1-amd64.deb" && \
+    dpkg -i pandoc-2.2.2.1-1-amd64.deb && \
+    rm pandoc-2.2.2.1-1-amd64.deb
 
 ENV CONDA_DIR="/opt/miniconda-latest" \
     PATH="/opt/miniconda-latest/bin:$PATH"
