### 2019-09-13 2673013
Fix Dockerfile and use actual NiBS version in boilerplate.
