### 2019-09-13 ea47be1
Add fake fir_delays to fix test.
