### 2019-09-13 #1295 @Kevin-Mattheus-Moerman
@jdkent It is okay that the Zenodo archived paper is different from this one. The purpose of the archive is to capture the software. Ping me here when you've implemented the changes.
