### 2019-09-17 c8b9106
respond to suggested edits - expand affiliations - change several awkward phrases - add funding sources for neurohackademy and michelle voss
