+++ .joss/paper.md
@@ -14,9 +14,9 @@ authors:
     orcid: 0000-0002-9840-6257
     affiliation: 2
 affiliations:
- - name: Neuroscience Program, University of Iowa
+ - name: Interdisciplinary Graduate Program in Neuroscience, University of Iowa, Iowa City, United states of America
    index: 1
- - name: Montréal Neurological Institute
+ - name: Montréal Neurological Institute, McGill University Health Centre, Montreal, Quebec, Canada
    index: 2
 date: 30 January 2019
 bibliography: paper.bib
@@ -27,13 +27,13 @@ bibliography: paper.bib
 You may not be aware of it, but your brain is orchestrating a complex ballet of activity while reading this sentence.
 Whether it is following a dot or reading a sentence; the brain is evaluating input and sending motoric output to perform optimally/efficiently.
 We can measure this flurry of activity using functional Magnetic Resonance Imaging (fMRI).
-Traditional fMRI analysis emphasizes what regions are "activated/deactivated" during a task, but does not provide information on which regions are acting in synchrony or are being segregated.
+Traditional fMRI analysis emphasizes what regions are "activated/deactivated" during a task, but it does not provide information about which regions are acting in synchrony or are being segregated.
 Knowing the synchronous/segregated brain regions during a task gives insights on the potential organization of the brain.
-**NiBetaSeries seeks to inform about the organization of the brain by correlating activation/deactivation patterns between brain regions during a task.**
+**NiBetaSeries seeks to provide information about the organization of the brain by correlating activation/deactivation patterns between brain regions during a task.**
 
 To understand NiBetaSeries we need to answer two questions: what is a "beta" (or parameter estimate) and how can we analyze a series of betas?
 We have already mentioned betas by another name, activation/deactivation.
-The term beta comes from its use in the General Linear Model (GLM), an extension of linear regression.
+The term beta comes from its use in the General Linear Model (GLM), which is an extension of linear regression.
 fMRI signal evoked by a stimulus follows a relatively stereotyped shape from the Blood Oxygen Level Dependent (BOLD) response, which is reasonably modeled with a double-gamma function.
 The overall amplitude (i.e., activation/deactivation) of the double gamma function is determined by the beta coefficient.
 Thus larger betas mean greater activation and smaller or negative betas mean less activation or deactivation relative to a baseline.
@@ -43,18 +43,18 @@ NiBetaSeries, on the other hand, gives each trial its own beta estimate treating
 Two common methods for deriving single trial beta estimates are least squares all (LSA) and least squares separate (LSS) [@Mumford2012;@Turner2012a].
 LSA places all the trials in the same GLM, where each trial is a separate predictor.
 LSA works well when the trials are far apart in time because the BOLD response takes a long time to return to baseline.
-When the trials are close together, however, the bold responses start to overlap and the GLM cannot accurately attribute the variance in the fMRI data to trial 1 or trial 2, leading to unreliable beta estimates.
+When the trials are close together, however, the bold responses start to overlap and the GLM cannot accurately attribute the variance in the fMRI data to a single trial, leading to unreliable beta estimates.
 LSS tackles this problem by making as many GLMs as there are trials.
 For each trial, a GLM is created with two predictors: one is the trial of interest and the second is the combination of all the other trials.
 LSS reduces the amount of overlap (or more accurately correlation) between predictors, leading to more reliable individual beta estimates.
-NiBetaSeries currently implements LSS making it a more reasonable analysis choice for
-experiments with trials that occur closer together (e.g., 3-7 seconds apart on average).
-The output of LSA or LSS is a beta series for each voxel in our dataset.
[…]
