### 2019-09-17 #1295 @jdkent
@arokem could you check this [pull request](https://github.com/HBClab/NiBetaSeries/pull/206/files) to make sure I covered the above points? Thanks!
