### 2019-09-18 #1295 @jdkent
@Kevin-Mattheus-Moerman

From [this commit](https://github.com/HBClab/NiBetaSeries/pull/206/commits/c8b91067dbe78354ea1ccd7da21a286b9ae287ba):
- I changed a couple of the sentences to avoid awkward phrasing and the sentence missing "be".
- I added city/country to affiliations and I made my affiliation more explicit
- I added and verified links to funding sources.

Let me know if I've covered all your suggestions adequately.
