### 2019-09-19 #1295 @jdkent
@snastase - thanks again for your review!

@labarba & @Kevin-Mattheus-Moerman - thank you for helping making the text clearer and answering my questions!

@arokem - thank you for your guidance during this process!

This was a great experience!
